### 2025-05-12 #706 @b-rodrigues
rixpress: Reproducible Analytical Pipelines with Nix
Submitting Author Name: Bruno Rodrigues
Submitting Author Github Handle: @b-rodrigues
Repository:  https://github.com/b-rodrigues/rixpress
Version submitted: 0.2.0
Submission type: Standard

Version accepted: TBD

---

```
Package: rixpress
Title: Build Reproducible Analytical Pipelines With Nix
Version: 0.2.0
    person("Bruno", "Rodrigues", , "bruno@brodrigues.co", role = c("aut", "cre"))
Description: Streamlines the creation of reproducible analytical pipelines using
  `default.nix` expressions generated via `{rix}` for reproducibility. Define
  derivations in R or Python, chain them into a composition of pure functions
  and build the resulting pipeline using `Nix` as the underlying end-to-end build
  tool. Functions to plot a DAG representation of the pipeline are included,
  as well as functions to load and inspect intermediary results for interactive
  analysis. User experience heavily inspired by the `{targets}` package.
    igraph,
    processx
    ggdag,
    mockery,
    reticulate,
    rix,
    usethis,
    visNetwork
```

This package is intended to help users set up reproducible pipelines using the Nix programming language for enhanced reproducibility.

The target audience is anyone wanting to switch from "script-based workflows" to build automation. rixpress generates valid Nix expressions from simple R function to define reproducible pipelines, and is heavily inspired by {targets}. The main difference between {targets} and this package is that the "heavy lifting" is performed by Nix, and it works very closely with my previous packages called {rix} which allows data scientists to set up reproducible environments using Nix. Also, because the underlying engine is Nix, it is language-agnostic, and so it is possible to define steps that use Python. These steps written in Python are not executed with {reticulate}, but instead run in a dedicated Python environment. Data transfer between Python an R is facilitated with {reticulate} though.

The main inspiration of this packages is `{targets}` and in combination with `{rix}`, one could set up a pipeline in a reproducible environment as well.

Link to presubmission: https://github.com/ropensci/software-review/issues/699

@maurolepore

Because this package relies heavily on side effects, unit tests are quite cumbersome to write, so I set up this other repository: https://github.com/b-rodrigues/rixpress_demos which contains many example pipelines that run on each push to `{rixpress}`'s repository. Thanks to LLM's I was able to improve test coverage to 67% (see https://github.com/b-rodrigues/rixpress/actions/runs/14971090564)
