+++ man/rxp_make.Rd
@@ -21,3 +21,12 @@ A character vector of paths to the built outputs.
 \description{
 Runs \code{nix-build} with a quiet flag, outputting to \verb{_rixpress/result}.
 }
+\examples{
+\dontrun{
+  # Build the pipeline with default settings
+  rxp_make()
+  
+  # Build with verbose output and parallel execution
+  rxp_make(verbose = TRUE, max_jobs = 4, cores = 2)
+}
+}
