+++ NAMESPACE
@@ -6,7 +6,6 @@ export(adjust_import)
 export(dag_for_ci)
 export(export_nix_archive)
 export(generate_dag)
-export(get_nodes_edges)
 export(import_nix_archive)
 export(rixpress)
 export(rxp_copy)
