+++ pkgdown.yml
@@ -9,7 +9,7 @@ articles:
   e-ci: e-ci.html
   f-cmdstanr: f-cmdstanr.html
   g-logs: g-logs.html
-last_built: 2025-05-12T12:56Z
+last_built: 2025-05-12T12:59Z
 urls:
   reference: https://b-rodrigues.github.io/rixpress/reference
   article: https://b-rodrigues.github.io/rixpress/articles
