### 2025-05-12 6cc0a4e
print.derivation example
