### 2025-05-12 8f525d4
sapply -> vapply
