### 2025-05-12 4525069
sapply -> vapply
