### 2025-05-12 0c0291d
sapply -> vapply
