### 2025-05-12 03d106e
sapply -> vapply
