### 2025-05-12 932bbfe
vapply -> sapply for additional files, as length can change
