### 2025-05-12 #706 @b-rodrigues
Ok, so I've implemented the changes, but for the unit test coverage. As explained, the package relies a lot on side-effects, so increasing to 75% will be quite difficult, especially because the functions that are not tested are those that would required build artifacts in the Nix store. Mocking that would be pain in the bottom.
As a compromise, I set up this repo: https://github.com/b-rodrigues/rixpress_demos
with complete pipelines that test these functions.

Would this be ok?
