+++ man/rxp_py_file.Rd
@@ -61,20 +61,14 @@ folder.
     path = "data/dataset.csv",
     read_function = "pandas.read_csv"
   )
-  
-  # Read a geospatial file with geopandas
-  rxp_py_file(
-    name = geopandas_data,
-    path = "data/boundaries.shp",
-    read_function = "geopandas.read_file",
-    copy_data_folder = TRUE
-  )
-  
-  # Read all CSV files in a directory using a lambda function
-  rxp_py_file(
-    name = all_csvs,
-    path = "data",
-    read_function = "lambda x: [pandas.read_csv(os.path.join(x, f)) for f in os.listdir(x) if f.endswith('.csv')]"
-  )
+
+# Read all CSV files in a directory using a
+# user defined function
+ rxp_py_file(
+  name = mtcars_py,
+  path = 'data',
+  read_function = "read_many_csvs",
+  copy_data_folder = TRUE
+)
 }
 }
