+++ R/rxp_ga.R
@@ -1,4 +1,5 @@
 #' Run a pipeline on GitHub Actions
+#' @family utilities
 #' @details This function puts a `.yaml` file inside the `.github/workflows/`
 #'   folder on the root of your project. This workflow file expects both
 #'   scripts generated by `rxp_init()`, `gen-env.R` and `gen-pipeline.R` to be
