+++ man/rxp_quarto.Rd
@@ -54,3 +54,13 @@ To include built derivations in the document,
   )
 }
 }
+\seealso{
+Other derivations: 
+\code{\link{rxp_file_common}()},
+\code{\link{rxp_py2r}()},
+\code{\link{rxp_py_file}()},
+\code{\link{rxp_r}()},
+\code{\link{rxp_r2py}()},
+\code{\link{rxp_rmd}()}
+}
+\concept{derivations}
