+++ man/rxp_py_file.Rd
@@ -72,3 +72,13 @@ folder.
 )
 }
 }
+\seealso{
+Other derivations: 
+\code{\link{rxp_file_common}()},
+\code{\link{rxp_py2r}()},
+\code{\link{rxp_quarto}()},
+\code{\link{rxp_r}()},
+\code{\link{rxp_r2py}()},
+\code{\link{rxp_rmd}()}
+}
+\concept{derivations}
