+++ man/rxp_r.Rd
@@ -81,3 +81,13 @@ paths.
   )
 }
 }
+\seealso{
+Other derivations: 
+\code{\link{rxp_file_common}()},
+\code{\link{rxp_py2r}()},
+\code{\link{rxp_py_file}()},
+\code{\link{rxp_quarto}()},
+\code{\link{rxp_r2py}()},
+\code{\link{rxp_rmd}()}
+}
+\concept{derivations}
