+++ man/rxp_ga.Rd
@@ -25,3 +25,4 @@ to give read and write permissions to the GitHub Actions bot.
   rxp_ga()
 }
 }
+\concept{utilities}
