+++ R/plot_dag.R
@@ -1,6 +1,6 @@
 #' @title Export DAG of pipeline and prepare it for rendering on CI
 
-#' @family visualisation functions
+#' @family ci utilities
 #' @description This function generates a DOT file representation of the
 #'   pipeline DAG, suitable for visualization, potentially on CI platforms. It
 #'   is called by `rxp_ga()`.
@@ -155,7 +155,7 @@ rxp_ggdag <- function(nodes_and_edges = get_nodes_edges()) {
 
 #' @title Create a Directed Acyclic Graph (DAG) representing the pipeline
 #'   using `{visNetwork}`
-#' @family visualisation
+#' @family visualisation functions
 #' @description Uses `{visNetwork}` to generate the plot. `{visNetwork}` is a
 #'   soft dependency of `{rixpress}` so you need to install it to use this
 #'   function.
