+++ man/rxp_init.Rd
@@ -39,6 +39,7 @@ Creates (overwriting if they already exist):
 Other utilities: 
 \code{\link{print.derivation}()},
 \code{\link{rxp_copy}()},
+\code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
 \code{\link{rxp_load}()},
 \code{\link{rxp_read}()}
