+++ man/rxp_read.Rd
@@ -47,6 +47,7 @@ Other utilities:
 \code{\link{print.derivation}()},
 \code{\link{rxp_copy}()},
 \code{\link{rxp_init}()},
+\code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
 \code{\link{rxp_load}()}
 }
