+++ man/rxp_file_common.Rd
@@ -55,3 +55,4 @@ Creates a Nix expression with shared logic for R and Python file reading.
   )
 }
 }
+\keyword{internal}
