+++ man/rxp_common_setup.Rd
@@ -36,3 +36,4 @@ based on the given parameters.
   )
 }
 }
+\keyword{internal}
