+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,27 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: b-rodrigues
+
+---
+
+**Review example pipelines**
+Take a look at the [rixpress_demos](https://github.com/b-rodrigues/rixpress_demos/tree/master) repository and find a pipeline that is close to what you’re trying to achieve. Maybe this could help
+you understand the issue you’re facing and describe what you want to achieve more easily.
+
+**Describe the bug**
+Give us a clear and concise description of what the bug is. Providing a reproducible example is very helpful. Please provide the default.nix that you’re using so that we can generate the same environment.
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - version of rixpress and rix being used
+
+**Additional context**
+Add any other context about the problem here.
