+++ R/rixpress.R
@@ -266,6 +266,7 @@ gen_flat_pipeline <- function(derivs) {
   types <- vapply(derivs, function(d) d$type, character(1))
   need_r <- get_need_r(types)
   need_py <- get_need_py(types)
+  need_jl <- get_need_jl(types)
 
   # Build function definitions
   function_defs <- ""
@@ -301,6 +302,22 @@ gen_flat_pipeline <- function(derivs) {
           cp ${pickleFile} $out
         '';
       };"
+    )
+  }
+  if (need_jl) {
+    function_defs <- paste0(
+      function_defs,
+      "\n  # Function to create Julia derivations
+  makeJlDerivation = { name, buildInputs, configurePhase, buildPhase, src ? null }:
+    defaultPkgs.stdenv.mkDerivation {
+      inherit name src;
+      dontUnpack = true;
+      buildInputs = buildInputs;
+      inherit configurePhase buildPhase;
+      installPhase = ''
+        cp ${name} $out/
+      '';
+    };"
     )
   }
 
@@ -398,6 +415,20 @@ gen_pipeline <- function(dag_file, flat_pipeline) {
           )
         }
       }
+    } else if (type == "rxp_jl") {
+      maker <- "makeJlDerivation"
+      script_cmd <- "julia -e \""
+      load_line <- function(dep, indent, unserialize_function) {
+        path <- paste0("${", dep, "}/", dep)
+        paste0(
+          dep,
+          " = ",
+          unserialize_function,
+          "(\"",
+          path,
+          "\")"
+        )
+      }
     } else {
       warning("Unsupported type for derivation ", deriv_name)
       next
@@ -436,7 +467,7 @@ gen_pipeline <- function(dag_file, flat_pipeline) {
     }
     script_idx <- build_phase_idx + script_idx[1]
 
-    # Determine indentation for R scripts, none for Python
+    # Determine indentation for R scripts, none for Python or Julia
     indent <- if (type == "rxp_r") {
       if (script_idx + 1 <= length(pipeline)) {
         sub("^([[:space:]]*).*", "\\1", pipeline[script_idx + 1])
@@ -485,7 +516,6 @@ generate_libraries_from_nix <- function(
   )
 }
 
-
 #' @noRd
 get_need_r <- function(types) {
   any(
@@ -497,4 +527,9 @@ get_need_r <- function(types) {
 #' @noRd
 get_need_py <- function(types) {
   any(types %in% c("rxp_py", "rxp_py_file"))
+}
+
+#' @noRd
+get_need_jl <- function(types) {
+  any(types %in% c("rxp_jl"))
 }
