+++ data-raw/jl_example/gen-pipeline.R
@@ -4,11 +4,15 @@ library(igraph)
 list(
   rxp_jl(d_size, '150'),
 
-  rxp_jl(
-    data,
-    '0.1randn(d_size,d_size) + reshape(cholesky(gridlaplacian(d_size,d_size) + 0.003I) \ randn(d_size*d_size), d_size, d_size)',
-    additional_files = "functions.jl"
-  ),
+rxp_jl(
+  data,
+  "0.1randn(d_size,d_size) + reshape( \
+     cholesky(gridlaplacian(d_size,d_size) + 0.003I) \\ randn(d_size*d_size), \
+     d_size, \
+     d_size \
+   )",
+  additional_files = "functions.jl"
+),
   rxp_jl(
     laplace_df,
     'DataFrame(data, :auto)',
