### 2025-05-21 c353f7f
done but must find a way to retrieve rayshader object
