### 2025-05-21 021c0ad
Refactor core pipeline generation and parsing logic
This commit introduces a significant refactoring of the R package's core functionality related to Nix pipeline generation, dependency parsing, and code quality.

Key changes include:

1.  **Generic Derivation Function (`R/rixpress.R`, `R/derivs.R`):**
    *   Introduced a generic `makeDerivation` Nix function in `pipeline.nix` generation, replacing separate `makeRDerivation`, `makePyDerivation`, and `makeJlDerivation`.
    *   Updated `rxp_*` functions in `R/derivs.R` to use this generic Nix function with a `type` parameter (e.g., "R", "Py", 
[…]
