+++ R/derivs.R
@@ -135,8 +135,9 @@ rxp_r <- function(
     src_snippet <- ""
   }
 
+  # Uses the generic makeDerivation Nix function with type = "R".
   snippet <- sprintf(
-    "  %s = makeRDerivation {\n    name = \"%s\";\n  %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
+    "  %s = makeDerivation {\n    type = \"R\";\n    name = \"%s\";\n  %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
     out_name,
     out_name,
     src_snippet,
@@ -302,9 +303,9 @@ rxp_py <- function(
     src_snippet <- ""
   }
 
-  # Generate the Nix snippet
+  # Generate the Nix snippet using the generic makeDerivation with type = "Py".
   snippet <- sprintf(
-    "  %s = makePyDerivation {\n    name = \"%s\";\n  %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
+    "  %s = makeDerivation {\n    type = \"Py\";\n    name = \"%s\";\n  %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
     out_name,
     out_name,
     src_snippet,
@@ -462,8 +463,7 @@ rxp_qmd <- function(
 #' @param nix_env Character, path to the Nix environment file.
 #' @param build_phase Character, the language-specific build phase script.
 #' @param type Character, the type of derivation ("rxp_r" or "rxp_py").
-#' @param derivation_func Character, the Nix derivation function
-#'   ("makeRDerivation" or "makePyDerivation").
+#' @param derivation_func Character, should now always be "makeDerivation".
 #' @param library_ext Character, the library file extension ("R" or "py").
 #' @param env_var List, defaults to NULL. A named list of environment variables
 #'   to set before running the script, e.g., c(DATA_PATH = "/path/to/data").
@@ -536,11 +536,15 @@ rxp_file_common <- function(
     build_phase <- paste0(env_exports, build_phase)
   }
 
-  # Build the Nix derivation snippet
+  # Determine Nix type parameter ("R" or "Py")
+  nix_type_param <- if (type == "rxp_r") "R" else if (type == "rxp_py") "Py" else stop("Unsupported type for rxp_file_common")
+
+  # Build the Nix derivation snippet using the generic makeDerivation.
+  # nix_type_param will be "R" or "Py" based on the calling function (rxp_r_file or rxp_py_file).
   snippet <- sprintf(
-    "  %s = %s {\n    name = \"%s\";\n    src = %s;\n    buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
+    "  %s = makeDerivation {\n    type = \"%s\";\n    name = \"%s\";\n    src = %s;\n    buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
     out_name,
-    derivation_func,
+    nix_type_param,
     out_name,
     src_part,
     base,
@@ -551,7 +555,7 @@ rxp_file_common <- function(
   list(
     name = out_name,
     snippet = snippet,
-    type = type,
+    type = type, # This is the internal R type, e.g. "rxp_r_file"
[…]
