+++ R/parse_pkgs.R
@@ -2,9 +2,9 @@
 #'
 #' @param nix_file Defaults to "default.nix", path to the default.nix file
 #' @param project_path Path to root of project, typically "."
-#' @param block_name Name of the block to parse (e.g., "rpkgs" or "pyconf")
-#' @param transform Function to transform package names (default: identity)
-#' @return List of package names, or NULL if the block is not found
+#' @param block_name Name of the block to parse (e.g., "rpkgs", "pyconf", "jlconf").
+#' @param transform Function to transform package names (default: identity).
+#' @return List of package names, or NULL if the block is not found.
 #' @noRd
 parse_packages <- function(
   nix_file,
@@ -14,16 +14,19 @@ parse_packages <- function(
 ) {
   lines <- readLines(file.path(project_path, nix_file))
 
-  # Find the start of the block
-  # (e.g., "rpkgs = builtins.attrValues {" or "pyconf = ...")
+  # Regex to find the start of the specified package block.
+  # It handles two main patterns:
+  # 1.  `<block_name> = builtins.attrValues { ... }` (common for R, Python).
+  # 2.  `<block_name> = pkgs.julia<...>.withPackages [ ... ]` (common for Julia).
+  # The `(?: ... )` creates a non-capturing group for the OR condition.
   start_pattern <- paste0(
     "^\\s*",
     block_name,
-    "\\s*=\\s*(?:",
-    "builtins\\.attrValues\\s*\\{", # matches rpkgs (or pyconf) = builtins.attrValues {
-    "|",                            # or, matches Julia blocks
-    "pkgs\\.julia(?:[-_\\.][A-Za-z0-9]+)*\\.withPackages\\s*\\[",
-    ")"
+    "\\s*=\\s*(?:", # Start of non-capturing group for OR
+    "builtins\\.attrValues\\s*\\{", # Pattern 1: R/Python style
+    "|",                            # OR
+    "pkgs\\.julia(?:[-_\\.][A-Za-z0-9]+)*\\.withPackages\\s*\\[", # Pattern 2: Julia style
+    ")"                             # End of non-capturing group
   )
 
   start_idx <- grep(start_pattern, lines)
@@ -43,14 +46,16 @@ parse_packages <- function(
   block_lines <- lines[(start_idx + 1):(end_idx - 1)]
   block_lines <- gsub("#.*", "", block_lines)
   block_lines <- trimws(block_lines)
-  block_lines <- block_lines[block_lines != ""]
+  block_lines <- block_lines[block_lines != ""] # Remove empty lines
 
-  # Remove inherit statements
-  # (e.g., "inherit (pkgs.rPackages)" or "inherit (pkgs.python312Packages)")
-  inherit_pattern <- "inherit \\(pkgs\\.[a-zA-Z0-9]+\\)"
+  # Remove `inherit (...)` statements as they don't list actual package names
+  # directly but rather groups of packages from Nixpkgs.
+  # e.g., "inherit (pkgs.rPackages);" or "inherit (pkgs.python312Packages);".
+  # The actual package names are expected to be listed individually if used.
+  inherit_pattern <- "inherit\\s*\\(pkgs\\.[a-zA-Z0-9]+\\)(?:\\s*;)?|inherit\\s*pkgs\\s*;"
   block_lines <- gsub(inherit_pattern, "", block_lines)
 
-  block_lines <- gsub(";", "", block_lines)
+  block_lines <- gsub(";", "", block_lines) # Remove any remaining semicolons
 
   # Split into package names and apply transformation
   packages <- unlist(strsplit(paste(block_lines, collapse = " "), "\\s+"))
@@ -75,10 +80,12 @@ parse_rpkgs_git <- function(
 
[…]

[…]
