+++ R/rixpress.R
@@ -86,7 +86,9 @@
 
   combined_additional_files <- lapply(unique_nix_envs, function(env) {
     files_for_this_env <- unlist(all_add_files[all_nix_envs == env])
-    unique_files <- unique(files_for_this_env[!is.na(files_for_this_env) & nzchar(files_for_this_env)])
+    unique_files <- unique(files_for_this_env[
+      !is.na(files_for_this_env) & nzchar(files_for_this_env)
+    ])
     if (length(unique_files) == 0) {
       return("")
     }
@@ -103,14 +105,20 @@
 # It iterates over the unique Nix environment configurations provided by
 # `.process_nix_configs` and calls `generate_libraries_from_nix` for each,
 # which creates the `libraries.[R|py|jl]` scripts in the `_rixpress` directory.
-.orchestrate_library_generation <- function(processed_nix_configs, project_path) {
-  suppressWarnings({ # Warnings are suppressed as generate_libraries_from_nix might warn if no packages are found.
+.orchestrate_library_generation <- function(
+  processed_nix_configs,
+  project_path
+) {
+  suppressWarnings({
+    # Warnings are suppressed as generate_libraries_from_nix might warn if no packages are found.
     for (i in seq_along(processed_nix_configs$unique_env)) {
       current_env <- processed_nix_configs$unique_env[i]
       current_files_list <- processed_nix_configs$additional_files_combined[[i]]
 
       # Ensure current_files is a character vector, not a list containing a single empty string
-      actual_files_param <- if (length(current_files_list) == 1 && current_files_list == "") {
+      actual_files_param <- if (
+        length(current_files_list) == 1 && current_files_list == ""
+      ) {
         ""
       } else {
         current_files_list
@@ -189,11 +197,18 @@ parse_nix_envs <- function(derivs) {
   derivs <- lapply(
     derivs,
     function(d) {
-      host_scripts <- list.files("_rixpress", pattern = d$base_name, full.names = FALSE) # Get just filenames
+      host_scripts <- list.files(
+        "_rixpress",
+        pattern = d$base_name,
+        full.names = FALSE
+      ) # Get just filenames
       if (length(host_scripts) == 0) {
         warning(paste0(
-          "No library script found in '_rixpress' for Nix env base '", d$base_name,
-          "' (derived from '", d$nix_env, "'). ",
+          "No library script found in '_rixpress' for Nix env base '",
+          d$base_name,
+          "' (derived from '",
+          d$nix_env,
+          "'). ",
           "This might be an issue if this environment is expected to provide libraries."
         ))
         # Allow to proceed but sandbox target name will be based on an empty string if no script.
@@ -203,14 +218,22 @@ parse_nix_envs <- function(derivs) {
       } else {
         if (length(host_scripts) > 1) {
           warning(paste0(
-            "Multiple library scripts found for Nix env base '", d$base_name,
-            "': ", paste(host_scripts, collapse = ", "), ". Using the first one: ", host_scripts[1]
+            "Multiple library scripts found for Nix env base '",
+            d$base_name,
+            "': ",
+            paste(host_scripts, collapse = ", "),
+            ". Using the first one: ",
+            host_scripts[1]
           ))
         }
         d$host_library_script_name <- host_scripts[1]
       }
       # The target name in the sandbox is derived from the host script name.
[…]
