+++ NAMESPACE
@@ -13,6 +13,7 @@ export(rxp_ga)
 export(rxp_ggdag)
 export(rxp_init)
 export(rxp_inspect)
+export(rxp_jl)
 export(rxp_list_logs)
 export(rxp_load)
 export(rxp_make)
