+++ man/rxp_r.Rd
@@ -83,6 +83,7 @@ paths.
 }
 \seealso{
 Other derivations: 
+\code{\link{rxp_jl}()},
 \code{\link{rxp_py}()},
 \code{\link{rxp_py_file}()},
 \code{\link{rxp_qmd}()},
