+++ man/rxp_qmd.Rd
@@ -56,6 +56,7 @@ To include built derivations in the document,
 }
 \seealso{
 Other derivations: 
+\code{\link{rxp_jl}()},
 \code{\link{rxp_py}()},
 \code{\link{rxp_py_file}()},
 \code{\link{rxp_r}()},
