+++ man/rxp_py_file.Rd
@@ -74,6 +74,7 @@ folder.
 }
 \seealso{
 Other derivations: 
+\code{\link{rxp_jl}()},
 \code{\link{rxp_py}()},
 \code{\link{rxp_qmd}()},
 \code{\link{rxp_r}()},
