+++ R/derivs.R
@@ -1,3 +1,49 @@
+#' Helper function to generate a Nix derivation snippet
+#' 
+#' @param out_name Character, name of the derivation
+#' @param src_snippet Character, the src part of the derivation
+#' @param base Character, base name for buildInputs and configurePhase
+#' @param build_phase Character, the build phase commands
+#' @param derivation_type Character, one of "R", "Py", "Jl", "Qmd", "Rmd"
+#' @return Character string with the formatted Nix derivation
+#' @keywords internal
+make_derivation_snippet <- function(
+  out_name,
+  src_snippet,
+  base,
+  build_phase,
+  derivation_type
+) {
+  # Determine the derivation function based on type
+  derivation_func <- switch(derivation_type,
+    "R" = "makeRDerivation",
+    "Py" = "makePyDerivation",
+    "Jl" = "makeJlDerivation",
+    "Qmd" = "defaultPkgs.stdenv.mkDerivation",
+    "Rmd" = "defaultPkgs.stdenv.mkDerivation",
+    stop("Unknown derivation type: ", derivation_type)
+  )
+  
+  # Format the build phase with appropriate indentation
+  formatted_build_phase <- if(derivation_type %in% c("Qmd", "Rmd")) {
+    paste0("\n", build_phase, "\n    ")
+  } else {
+    paste0("\n      ", build_phase, "\n    ")
+  }
+  
+  # Generate the snippet
+  sprintf(
+    "  %s = %s {\n    name = \"%s\";\n    %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''%s'';\n  };",
+    out_name,
+    derivation_func,
+    out_name,
+    src_snippet,
+    base,
+    base,
+    formatted_build_phase
+  )
+}
+
 #' Create a Nix expression running an R function
 #' @family derivations
 #' @param name Symbol, name of the derivation.
@@ -128,21 +174,19 @@ rxp_r <- function(
   if (length(fileset_parts) > 0) {
     fileset_nix <- paste0("./", fileset_parts, collapse = " ")
     src_snippet <- sprintf(
-      "   src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n ",
+      "src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n",
       fileset_nix
     )
   } else {
     src_snippet <- ""
   }
 
-  snippet <- sprintf(
-    "  %s = makeRDerivation {\n    name = \"%s\";\n  %s  buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''\n      %s\n    '';\n  };",
-    out_name,
-    out_name,
-    src_snippet,
-    base,
-    base,
-    build_phase
+  snippet <- make_derivation_snippet(
+    out_name = out_name,
+    src_snippet = src_snippet,
+    base = base,
+    build_phase = build_phase,
+    derivation_type = "R"
   )
 
   list(
@@ -295,7 +339,7 @@ rxp_py <- function(
   if (length(fileset_parts) > 0) {
     fileset_nix <- paste0("./", fileset_parts, collapse = " ")
     src_snippet <- sprintf(
-      "   src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n ",
[…]
