+++ R/derivs.R
@@ -1,5 +1,5 @@
 #' Helper function to generate a Nix derivation snippet
-#' 
+#'
 #' @param out_name Character, name of the derivation
 #' @param src_snippet Character, the src part of the derivation
 #' @param base Character, base name for buildInputs and configurePhase
@@ -15,7 +15,8 @@ make_derivation_snippet <- function(
   derivation_type
 ) {
   # Determine the derivation function based on type
-  derivation_func <- switch(derivation_type,
+  derivation_func <- switch(
+    derivation_type,
     "R" = "makeRDerivation",
     "Py" = "makePyDerivation",
     "Jl" = "makeJlDerivation",
@@ -23,14 +24,14 @@ make_derivation_snippet <- function(
     "Rmd" = "defaultPkgs.stdenv.mkDerivation",
     stop("Unknown derivation type: ", derivation_type)
   )
-  
+
   # Format the build phase with appropriate indentation
-  formatted_build_phase <- if(derivation_type %in% c("Qmd", "Rmd")) {
+  formatted_build_phase <- if (derivation_type %in% c("Qmd", "Rmd")) {
     paste0("\n", build_phase, "\n    ")
   } else {
     paste0("\n      ", build_phase, "\n    ")
   }
-  
+
   # Generate the snippet
   sprintf(
     "  %s = %s {\n    name = \"%s\";\n%s    buildInputs = %sBuildInputs;\n    configurePhase = %sConfigurePhase;\n    buildPhase = ''%s'';\n  };",
@@ -174,7 +175,7 @@ rxp_r <- function(
   if (length(fileset_parts) > 0) {
     fileset_nix <- paste0("./", fileset_parts, collapse = " ")
     src_snippet <- sprintf(
-      "     src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n   ",
+      "     src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n",
       fileset_nix
     )
   } else {
@@ -339,7 +340,7 @@ rxp_py <- function(
   if (length(fileset_parts) > 0) {
     fileset_nix <- paste0("./", fileset_parts, collapse = " ")
     src_snippet <- sprintf(
-      "     src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n   ",
+      "     src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n",
       fileset_nix
     )
   } else {
@@ -473,7 +474,10 @@ rxp_qmd <- function(
   # Generate the Nix derivation snippet with updated buildInputs and configurePhase
   snippet <- make_derivation_snippet(
     out_name = out_name,
-    src_snippet = sprintf("    src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n", fileset_nix),
+    src_snippet = sprintf(
+      "    src = defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    };\n",
+      fileset_nix
+    ),
     base = base,
     build_phase = build_phase,
     derivation_type = "Qmd"
@@ -582,7 +586,7 @@ rxp_file_common <- function(
     src_snippet = sprintf("    src = %s;\n", src_part),
     base = base,
[…]
