+++ man/rxp_py.Rd
@@ -36,7 +36,8 @@ the default behavior uses \code{pickle.dump}. Define this function in
 Python function used to unserialize the object. It must accept one
 argument: the file path.}
 
-\item{env_var}{Character vector, defaults to NULL. A named vector of environment variables
+\item{env_var}{Character vector, defaults to NULL. A named vector of
+environment variables
 before running the Python script, e.g., c(PYTHONPATH = "/path/to/modules").
 Each entry will be added as an export statement in the build phase.}
 }
