+++ README.md
@@ -68,12 +68,12 @@ list(
 
 Running `rixpress()` generates a `pipeline.nix` file, which contains all
 the build instructions of all the derivations and final outputs as `Nix`
-code. It is possible to define derivations that run Python code as well,
-and objects can be passed to and from R or Python by using `rxp_py2r()`
-and `ryp_r2py()`, or by using a common serialization format such as
-JSON. By default, calling `rixpress()` also builds the pipeline, but
-it’s possible to only generate the `pipeline.nix` file and then build
-the pipeline at a later stage using:
+code. It is possible to define derivations that run Python or Julia code
+as well, and objects can be passed to and from R or Python by using
+`rxp_py2r()` and `ryp_r2py()`, or by using a common serialization format
+such as JSON. By default, calling `rixpress()` also builds the pipeline,
+but it’s possible to only generate the `pipeline.nix` file and then
+build the pipeline at a later stage using:
 
     rxp_make()
 
@@ -81,8 +81,9 @@ The build process assumes the presence of a `default.nix` file which
 defines the computational environment the pipeline runs in; this
 `default.nix` file can be generated by using the `{rix}` package. This
 `default.nix` defines an environment with R, R packages (and also Python
-and Python packages if needed), Quarto, and all required system-level
-dependencies pinned to a specific date to ensure reproducibility.
+and/or Julia and their packages if needed), Quarto, and all required
+system-level dependencies pinned to a specific date to ensure
+reproducibility.
 
 In the example above, the first derivation loads `mtcars.csv` (it
 actually should be a `.psv` file, since the data is separated by pipes
