### 2025-05-22 2693ea4
refactor generate_dag.R
