+++ R/generate_dag.R
@@ -16,147 +16,37 @@
 #' }
 #' @export
 generate_dag <- function(rxp_list, output_file = "_rixpress/dag.json") {
+  # Validate inputs
+  if (!is.list(rxp_list)) {
+    stop("rxp_list must be a list of derivations")
+  }
+
+  # Create output directory if it doesn't exist
   dir.create(dirname(output_file), recursive = TRUE, showWarnings = FALSE)
 
+  # Extract all derivation names for dependency checking
   all_derivs_names <- sapply(rxp_list, function(x) (x$name))
 
   n <- length(rxp_list)
   dag <- vector("list", n)
   defined <- character(n)
 
+  # Process each derivation
   for (i in seq_along(rxp_list)) {
-    d <- rxp_list[[i]]
-    name <- d$name
-    type <- d$type
-    unserialize_function <- d$unserialize_function
+    deriv <- rxp_list[[i]]
+    name <- deriv$name
+    type <- deriv$type
+    unserialize_function <- deriv$unserialize_function
 
-    if (type == "rxp_r" || type == "rxp_py2r" || type == "rxp_r2py") {
-      snippet <- d$snippet
-      # Extract the content inside the Rscript -e quotes
-      # (allowing for multiple lines)
-      m <- regexec('Rscript -e \\"([\\s\\S]*?)\\"', snippet, perl = TRUE)
-      match <- regmatches(snippet, m)
-
-      derivs_to_consider <- Filter(
-        function(x) `!=`(x, name),
-        all_derivs_names
-      )
-
-      deps <- sapply(
-        derivs_to_consider,
-        function(name) any(grepl(paste0("\\b", name, "\\b"), unlist(match)))
-      )
-
-      # Only keep deps
-      deps <- Filter(isTRUE, deps)
-      deps <- names(deps)
-    } else if (type == "rxp_qmd" || type == "rxp_rmd") {
-      # Determine file path and extension based on type
-      if (type == "rxp_qmd") {
-        # Try both .qmd and .Qmd extensions
-        doc_file <- d$qmd_file
-        if (!file.exists(doc_file)) {
-          doc_file <- paste0(name, ".Qmd")
-          if (!file.exists(doc_file)) {
-            stop("Quarto file not found: ", name, ".qmd or ", name, ".Qmd")
-          }
-        }
-      } else {
-        # rxp_rmd
-        # Try .rmd and .Rmd extensions
-        doc_file <- d$rmd_file
-        if (!file.exists(doc_file)) {
-          doc_file <- paste0(name, ".Rmd")
-          if (!file.exists(doc_file)) {
-            stop("R Markdown file not found: ", name, ".rmd or ", name, ".Rmd")
-          }
-        }
-      }
-
-      # Read the document file
-      doc_content <- readLines(doc_file, warn = FALSE)
-      doc_text <- paste(doc_content, collapse = "\n")
-
-      # Extract R code chunks (between ```{r} and ```)
-      chunk_pattern <- "```\\{r\\}[\\s\\S]*?```"
-      chunks <- regmatches(
-        doc_text,
-        gregexpr(chunk_pattern, doc_text, perl = TRUE)
-      )[[1]]
-
-      # Process each chunk to find rxp_read() or rxp_load() calls
-      deps <- character(0)
-      for (chunk in chunks) {
-        # Remove the ```{r} and ``` delimiters
-        code <- sub("```\\{r\\}\\s*", "", chunk)
-        code <- sub("```\\s*$", "", code)
-
-        # Match rxp_read("name") or rxp_load("name"), with or without rixpress::
-        pattern <- "(rixpress::)?rxp_(read|load)\\s*\\(\\s*['\"](\\w+)['\"]\\s*\\)"
[…]
