### 2025-05-22 568e5ef
better handling of dependencies
