### 2025-05-22 8ad2404
refactor parse_pkgs.R
