+++ R/parse_pkgs.R
@@ -1,6 +1,6 @@
 #' Extract packages from a specified block in a default.nix file
 #'
-#' @param nix_file Defaults to "default.nix", path to the default.nix file
+#' @param nix_file Path to the default.nix file
 #' @param project_path Path to root of project, typically "."
 #' @param block_name Name of the block to parse (e.g., "rpkgs" or "pyconf")
 #' @param transform Function to transform package names (default: identity)
@@ -12,7 +12,13 @@ parse_packages <- function(
   block_name,
   transform = identity
 ) {
-  lines <- readLines(file.path(project_path, nix_file))
+  file_path <- file.path(project_path, nix_file)
+  if (!file.exists(file_path)) {
+    warning(paste("File not found:", file_path))
+    return(NULL)
+  }
+
+  lines <- readLines(file_path)
 
   # Find the start of the block
   # (e.g., "rpkgs = builtins.attrValues {" or "pyconf = ...")
@@ -21,7 +27,7 @@ parse_packages <- function(
     block_name,
     "\\s*=\\s*(?:",
     "builtins\\.attrValues\\s*\\{", # matches rpkgs (or pyconf) = builtins.attrValues {
-    "|",                            # or, matches Julia blocks
+    "|", # or, matches Julia blocks
     "pkgs\\.julia(?:[-_\\.][A-Za-z0-9]+)*\\.withPackages\\s*\\[",
     ")"
   )
@@ -34,23 +40,23 @@ parse_packages <- function(
 
   # Find the end of the block ("};" or "];")
   end_idxs <- grep("^\\s*(\\};|\\];)", lines, perl = TRUE)
-  end_idx  <- end_idxs[end_idxs > start_idx][1]
+  end_idx <- end_idxs[end_idxs > start_idx][1]
   if (is.na(end_idx)) {
     stop(paste("Could not find the end of the", block_name, "block"))
   }
 
   # Extract and clean lines within the block
   block_lines <- lines[(start_idx + 1):(end_idx - 1)]
-  block_lines <- gsub("#.*", "", block_lines)
-  block_lines <- trimws(block_lines)
-  block_lines <- block_lines[block_lines != ""]
+  block_lines <- gsub("#.*", "", block_lines) # Remove comments
+  block_lines <- trimws(block_lines) # Remove whitespace
+  block_lines <- block_lines[block_lines != ""] # Remove empty lines
 
   # Remove inherit statements
   # (e.g., "inherit (pkgs.rPackages)" or "inherit (pkgs.python312Packages)")
   inherit_pattern <- "inherit \\(pkgs\\.[a-zA-Z0-9]+\\)"
   block_lines <- gsub(inherit_pattern, "", block_lines)
 
-  block_lines <- gsub(";", "", block_lines)
+  block_lines <- gsub(";", "", block_lines) # Remove semicolons
 
   # Split into package names and apply transformation
   packages <- unlist(strsplit(paste(block_lines, collapse = " "), "\\s+"))
@@ -71,14 +77,17 @@ parse_rpkgs_git <- function(
   project_path,
   transform = identity
 ) {
-  lines <- readLines(file.path(project_path, nix_file))
+  file_path <- file.path(project_path, nix_file)
+  if (!file.exists(file_path)) {
+    warning(paste("File not found:", file_path))
+    return(NULL)
+  }
 
+  lines <- readLines(file_path)
   git_packages <- c()
 
   # Pattern to find the start of a buildRPackage definition
   # e.g., my_pkg = (pkgs.rPackages.buildRPackage {
-  # We capture the variable name for context if needed,
-  # but don't use it for the package name itself
[…]
