### 2025-05-22 9c9d4e3
Fix function exports and documentation
