+++ man/import_formatter_jl.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/parse_pkgs.R
+\name{import_formatter_jl}
+\alias{import_formatter_jl}
+\title{Format Julia import statements}
+\usage{
+import_formatter_jl(package)
+}
+\arguments{
+\item{package}{Character, name of the package}
+}
+\value{
+Character string with import statement
+}
+\description{
+Format Julia import statements
+}
