+++ man/extract_julia_dependencies.Rd
@@ -1,0 +1,21 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/generate_dag.R
+\name{extract_julia_dependencies}
+\alias{extract_julia_dependencies}
+\title{Extract dependencies from Julia script derivations}
+\usage{
+extract_julia_dependencies(deriv, name, all_derivs_names)
+}
+\arguments{
+\item{deriv}{The derivation object}
+
+\item{name}{The name of the derivation}
+
+\item{all_derivs_names}{All derivation names}
+}
+\value{
+A character vector of dependency names
+}
+\description{
+Extract dependencies from Julia script derivations
+}
