+++ man/gen_flat_pipeline.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rixpress.R
+\name{gen_flat_pipeline}
+\alias{gen_flat_pipeline}
+\title{gen_flat_pipeline Internal function used to generate most of the boilerplate in pipeline.nix}
+\usage{
+gen_flat_pipeline(derivs)
+}
+\arguments{
+\item{derivs}{List of derivation objects}
+}
+\value{
+Character string with the pipeline
+}
+\description{
+gen_flat_pipeline Internal function used to generate most of the boilerplate in pipeline.nix
+}
