+++ R/generate_dag.R
@@ -45,7 +45,7 @@ generate_dag <- function(rxp_list, output_file = "_rixpress/dag.json") {
       type,
       name,
       all_derivs_names,
-      defined[1:(i-1)]  # Pass previously defined derivations for filtering
+      defined[1:(i - 1)] # Pass previously defined derivations for filtering
     )
     # Add the derivation to the DAG
     dag[[i]] <- list(
@@ -93,7 +93,7 @@ extract_dependencies <- function(
   } else {
     stop("Unknown derivation type: ", type)
   }
-  
+
   # Filter to only include dependencies that have been defined previously
   # This ensures the DAG remains acyclic
   if (length(defined_derivs) > 0) {
