+++ data-raw/jl_example/functions.R
@@ -1,4 +1,4 @@
-prepare_data <- function(laplace){
+prepare_data <- function(laplace) {
   laplace_df |>
     mutate(
       x_id = row_number()
@@ -10,16 +10,16 @@ prepare_data <- function(laplace){
     )
 }
 
-make_gg <- function(laplace_long_df){
+make_gg <- function(laplace_long_df) {
   laplace_long_df |>
     ggplot(aes(x = x_id, y = y_id, z = z)) +
-    stat_summary_hex(fun = function(x) mean(x), bins = 45)  +
-    scale_fill_viridis_c(option = 12) + 
+    stat_summary_hex(fun = function(x) mean(x), bins = 45) +
+    scale_fill_viridis_c(option = 12) +
     theme_void() +
     theme(legend.position = "none") +
     labs(subtitle = "hexagonal 2-d heatmap of laplacian matrix")
 }
 
-save_gg <- function(path, gg){
+save_gg <- function(path, gg) {
   ggsave("gg.png", gg)
 }
