+++ man/extract_dependencies.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/generate_dag.R
+\name{extract_dependencies}
+\alias{extract_dependencies}
+\title{Extract dependencies for a derivation}
+\usage{
+extract_dependencies(deriv, type, name, all_derivs_names, defined_derivs)
+}
+\arguments{
+\item{deriv}{The derivation object}
+
+\item{type}{The type of derivation}
+
+\item{name}{The name of the derivation}
+
+\item{all_derivs_names}{All derivation names}
+
+\item{defined_derivs}{Previously defined derivations}
+}
+\value{
+A character vector of dependency names
+}
+\description{
+Extract dependencies for a derivation
+}
