+++ vignettes/d2-polyglot-julia.Rmd
@@ -1,0 +1,229 @@
+---
+title: "Polyglot pipelines with Julia and R"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{d2-polyglot-julia}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+This vignette demonstrates how to build a polyglot pipeline using R and Julia.
+It assumes you've read `vignette("b-core-functions")` and are familiar with the
+basic concepts of `{rixpress}`.
+
+For various examples of polyglot pipelines, check out the [rixpress_demos
+repository](https://github.com/b-rodrigues/rixpress_demos/).
+
+## Generating *waveshaders* data using Julia and plotting it using R
+
+This example was provided by Moritz Schauer (@mschauer).
+example adapted from: https://github.com/frankiethull/waveshaders/tree/main/experiments/simple_example
+
+`{rixpress}` makes it easy to write polyglot (multilingual) data science
+pipelines with derivations that run R or Julia code. This vignette explains how
+you can easily set up such a pipeline.
+
+### Setting up the environment
+
+Start your project by calling `rixpress::rxp_init()` and edit `gen-env.R`.
+You can open `gen-env.R` in your favourite text editor and define the execution
+environment there. For a Julia and R polyglot pipeline, you'll need to add Julia
+configuration:
+
+```{r, eval = FALSE}
+library(rix)
+
+rix(
+  date = "2025-05-19",
+  r_pkgs = c(
+    "arrow",
+    "dplyr",
+    "tidyr",
+    "ggplot2",
+    "hexbin"
+  ),
+  git_pkgs = list(
+    list(
+      package_name = "rix",
+      repo_url = "https://github.com/ropensci/rix/",
+      commit = "HEAD"
+    ),
+    list(
+      package_name = "rixpress",
+      repo_url = "https://github.com/b-rodrigues/rixpress",
+      commit = "HEAD"
+    )
+  ),
+  jl_conf = list(
+    jl_version = "1.10",
+    jl_pkgs = c(
+      "Arrow",
+      "DataFrames",
+      "SparseArrays",
+      "LinearAlgebra"
+    )
+  ),
+  ide = "none",
+  project_path = ".",
+  overwrite = TRUE
+)
+```
+
+Notice the `jl_conf` argument to `rix()`: this will install Julia and the listed
+Julia packages in that environment.
+
+Now that you've defined the execution environment of the pipeline, you can run
+the `gen-env.R` script, still from the temporary `Nix` shell by running
+`source("gen-env.R")`. This will generate the required `default.nix`. Then, quit
+R and the temporary shell (CTRL-D or `quit()` in R, `exit` in the terminal) and
+then build the environment defined by the freshly generated `default.nix` by
+typing `nix-build`. This will now build the execution environment of the
+pipeline. You can use this environment to work on your project interactively as
+usual. To learn more, check out [`{rix}`](https://docs.ropensci.org/rix/).
+
+### Creating the pipeline
+
+You can now edit the pipeline script in `gen-pipeline.R`. Here's an example of a
+pipeline that uses both R and Julia:
+
+```{r, eval = FALSE}
+library(rixpress)
+library(igraph)
+
+list(
+  rxp_jl(d_size, '150'),
+
+  rxp_jl(
+    data,
+    "0.1randn(d_size,d_size) + reshape( \
+    cholesky(gridlaplacian(d_size,d_size) + 0.003I) \\ randn(d_size*d_size), \
+    d_size, \
+    d_size \
+    )",
+    additional_files = "functions.jl"
[…]
