+++ vignettes/a-intro-concepts.Rmd
@@ -20,14 +20,15 @@ This vignette introduces key terminology and walks you through setting up a
 simple pipeline using `{rixpress}`. Understanding the concepts presented here is
 helpful but not mandatory. I recommend reading through this vignette, but don't
 worry if you don't grasp everything immediately. You can try building a simple
-pipeline by following the next vignette, `vignette("b-core-functions")`, and then
-return to this one; things should become clearer after some hands-on experience!
+pipeline by following the next vignette, `vignette("b-core-functions")`, and
+then return to this one; things should become clearer after some hands-on
+experience!
 
 Also, this package makes heavy use of its sister package `{rix}`, and I highly
-recommend that you first get familiar with `Nix` by using `{rix}`. With `{rix}`, you'll
-learn how to set up reproducible development environments for data science,
-which you can use for interactive data analysis work. Then, if you want to go
-one step further in the [reproducibility
+recommend that you first get familiar with `Nix` by using `{rix}`. With `{rix}`,
+you'll learn how to set up reproducible development environments for data
+science, which you can use for interactive data analysis work. Then, if you want
+to go one step further in the [reproducibility
 spectrum/continuum](https://www.researchgate.net/figure/Reproducibility-spectrum-as-Peng-2011-stated_fig1_354765302),
 this is where `{rixpress}` comes in.
 
