+++ R/derivs.R
@@ -1149,20 +1149,21 @@ print.derivation <- function(x, ...) {
 #' @return An object of class derivation which inherits from lists.
 #' @examples
 #' \dontrun{
-#'   # Basic usage, no custom serializer
-#'   rxp_jl(
-#'     name = filtered_df,
-#'     jl_expr = "filter(df, :col .> 10)"
-#'   )
+#' # Basic usage, no custom serializer
+#' rxp_jl(
+#'   name = filtered_df,
+#'   jl_expr = "filter(df, :col .> 10)"
+#' )
 #'
-#'   # Custom serialization: assume `save_my_obj(obj, path)` is defined in functions.jl
-#'   rxp_jl(
-#'     name = model_output,
-#'     jl_expr = "train_model(data)",
-#'     serialize_function = "save_my_obj",
-#'     additional_files = "functions.jl"
-#'   )
+#' # Custom serialization: assume `save_my_obj(obj, path)` is defined in functions.jl
+#' rxp_jl(
+#'   name = model_output,
+#'   jl_expr = "train_model(data)",
+#'   serialize_function = "save_my_obj",
+#'   additional_files = "functions.jl"
+#' )
 #' }
+#' @family derivations
 #' @export
 rxp_jl <- function(
   name,
