+++ man/rxp_jl.Rd
@@ -81,13 +81,5 @@ Other derivations:
 \code{\link{rxp_r}()},
 \code{\link{rxp_r_file}()},
 \code{\link{rxp_rmd}()}
-
-Other derivations: 
-\code{\link{rxp_py}()},
-\code{\link{rxp_py_file}()},
-\code{\link{rxp_qmd}()},
-\code{\link{rxp_r}()},
-\code{\link{rxp_r_file}()},
-\code{\link{rxp_rmd}()}
 }
 \concept{derivations}
