+++ man/extract_r_dependencies.Rd
@@ -19,3 +19,4 @@ A character vector of dependency names
 \description{
 Extract dependencies from R script derivations
 }
+\keyword{internal}
