+++ pkgdown.yml
@@ -1,4 +1,4 @@
-pandoc: 3.1.11.1
+pandoc: '3.6'
 pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
@@ -10,7 +10,7 @@ articles:
   e-ci: e-ci.html
   f-cmdstanr: f-cmdstanr.html
   g-logs: g-logs.html
-last_built: 2025-05-22T17:25Z
+last_built: 2025-05-23T09:48Z
 urls:
   reference: https://b-rodrigues.github.io/rixpress/reference
   article: https://b-rodrigues.github.io/rixpress/articles
