### 2025-05-23 6df35ba
refactor generate_dag, consider all deps
