### 2025-05-23 c41fcc7
Update extract_dependencies.Rd
