### 2025-07-10 641f102
handling of rxp_load in quarto documents
