### 2025-07-10 840c9ad
Update derivs.R
