### 2025-07-10 b18e42b
Update derivs.R
