### 2025-07-10 97e4f08
Update derivs.R
