### 2025-07-10 df3ede4
Update derivs.R
