### 2025-07-10 fe0a9a2
Update derivs.R
