+++ R/derivs.R
@@ -298,7 +298,8 @@ rxp_py <- function(
   nix_env = "default.nix",
   serialize_function = NULL,
   unserialize_function = NULL,
-  env_var = NULL
+  env_var = NULL,
+  user_functions = NULL # <-- NEW ARGUMENT
 ) {
   out_name <- deparse1(substitute(name))
   py_expr <- gsub("'", "\\'", py_expr, fixed = TRUE)
@@ -346,26 +347,66 @@ rxp_py <- function(
     if (env_exports != "") env_exports <- paste0(env_exports, "\n      ")
   }
 
-  fileset_parts <- setdiff(additional_files, "functions.py")
-  fileset_parts <- fileset_parts[nzchar(fileset_parts)]
-
-  # build copy command for additional files
-  copy_cmd <- ""
-  if (length(fileset_parts) > 0) {
-    copy_lines <- vapply(
+  # Prepare the fileset for src
+  # Combine additional_files and user_functions for the fileset
+  fileset_parts <- c()
+  if (!is.null(additional_files) && any(nzchar(additional_files))) {
+    fileset_parts <- c(
       fileset_parts,
+      additional_files[nzchar(additional_files)]
+    )
+  }
+  if (!is.null(user_functions) && any(nzchar(user_functions))) {
+    fileset_parts <- c(fileset_parts, user_functions[nzchar(user_functions)])
+  }
+
+  # build copy command for additional files (excluding user_functions)
+  copy_cmd <- ""
+  if (!is.null(additional_files) && any(nzchar(additional_files))) {
+    additional_files_clean <- additional_files[nzchar(additional_files)]
+    copy_lines <- vapply(
+      additional_files_clean,
       function(f) sprintf("cp -r ${./%s} %s", f, f),
       character(1)
     )
     copy_cmd <- paste0(paste(copy_lines, collapse = "\n      "), "\n      ")
   }
 
+  # build copy command for user_functions (explicit copy, not -r)
+  user_functions_copy_cmd <- ""
+  if (!is.null(user_functions) && any(nzchar(user_functions))) {
+    user_functions_clean <- user_functions[nzchar(user_functions)]
+    user_copy_lines <- vapply(
+      user_functions_clean,
+      function(f) sprintf("cp ${./%s} %s", f, f),
+      character(1)
+    )
+    user_functions_copy_cmd <- paste0(
+      paste(user_copy_lines, collapse = "\n      "),
+      "\n      "
+    )
+  }
+
+  # Generate import commands for user_functions
+  user_import_cmd <- ""
+  if (!is.null(user_functions) && any(nzchar(user_functions))) {
+    user_functions_clean <- user_functions[nzchar(user_functions)]
+    import_lines <- vapply(
+      user_functions_clean,
+      function(f) sprintf("exec(open('%s').read())", f),
+      character(1)
+    )
+    user_import_cmd <- paste0(paste(import_lines, collapse = "\n"), "\n")
+  }
+
   # Construct build_phase including cp commands then python execution
   build_phase <- paste0(
     env_exports,
     copy_cmd,
+    user_functions_copy_cmd,
     "python -c \"\n",
     "exec(open('libraries.py').read())\n",
+    user_import_cmd,
     "exec('",
     out_name,
     " = ",
@@ -380,7 +421,7 @@ rxp_py <- function(
   base <- gsub("[^a-zA-Z0-9]", "_", nix_env)
   base <- sub("_nix$", "", base)
 
-  # Prepare the src snippet only once using the same fileset_parts
+  # Prepare the src snippet with all files
   if (length(fileset_parts) > 0) {
     fileset_nix <- paste0("./", fileset_parts, collapse = " ")
[…]
