### 2025-07-21 #706 @wlandau
Bruno and I follow each other's work as members of the R community. We have not yet worked together directly on a project.

As the author of [`targets`](https://docs.ropensci.org/targets), I took a careful look at the [coi guidelines](https://devguide.ropensci.org/softwarereview_policies.html#coi):

> The potential editor or reviewer has a conflict of interest if:...The potential reviewer/editor has significantly contributed to a competitor project.

There is obvious overlap, but I would not say `rixpress` is a competitor. `rixpress` has a niche outside the scope of `targets`:

* [`nix-store`](https://nix.dev/manual/nix/latest/command-ref/nix-store.html) as the engine to run pipelines and store data.
* Polyglot pipelines where Python and Julia are first-class citizens alongside R.
* [Multi-environment pipelines](https://github.com/b-rodrigues/rixpress_demos/tree/master/r_multi_envs).

I checked with @ldecicco-USGS, who agreed.

---

## Review Comments

### Overview

`rixpress` is an excellent prospective addition to rOpenSci. It fills a valuable niche in reproducible computation, the engineering is fantastic, and the documentation is comprehensive. Because the quality is already so high, I did not need to spend much time checking package development minutia. I spent most of my review time on high-level issues and my own experience as a new user.

### Scope of this review

I reviewed:

* The documentation at <https://b-rodrigues.github.io/rixpress/>.
* Examples `basic_r`, `r_multi_envs`, and `r_py_xgboost` from https://github.com/b-rodrigues/rixpress_demos.
* The `rixpress` source code and test suite.

### Scope of `rixpress`

Arguably the most essential but most difficult part of developing any tool is establishing a clear and crisp set of requirements. Explicit pre-specified boundaries help prevent scope creep and ensures a package's priorities succeed long-term. For pipeline tools, scope is even more essential and more challenging than normal, both because of the [many different opinions about what a pipeline tool should do](https://github.com/pditommaso/awesome-pipeline), and because of the huge variety of pipelines users can create.

I developed `targets` as a highly opinionated tool with R-focused research-oriented scenarios in mind. This vision was somewhat implicit, and I did not have enough experience then to completely spell it out. I regularly hear from people who use it for cases I did not consider: simple ETL operations on big data, database query workflows, daily pipelines where historical runs matter, etc. Some users even approach `targets` as an Airflow-like tool rather than a Make-like one, and they are looking for a feature set closer to what [`maestro`](https://whipson.github.io/maestro/) provides.

You might have the same experience with `rixpress`. For example, users who switch from `targets` to `rixpress` may ask you to support [branching](https://books.ropensci.org/targets/static.html), [alternative DSLs](https://github.com/ropensci/targets/issues/1309), interactive debugging, fancy progress monitoring, alternative storage options, computing on clusters, alternative DAG visualizations, etc.

For `rixpress`, I would like to understand what the package may cover in the future, and what it definitely will not support. I think a dedicated section on scope in the documentation (possibly linked from the [issue templates](https://github.com/b-rodrigues/rixpress/tree/main/.github/ISSUE_TEMPLATE)) will help set expectations for users who request features, and it will help you maintain `rixpress` for years to come.

At this early stage, the main areas of focus seem to be as follows (please correct me if I am wrong):

1. Bringing pipeline functionality from [`nix-store`](https://nix.dev/manual/nix/latest/command-ref/nix-store.html) to R.
2. Interactive read-only inspection: visualization, reading from the data store, and historical runs.
3. R/Python/Julia interoperability.
4. Portability: through Nix itself, and continuous integration.

(1) seems like a promising direction because it invests in the qualities that makes `rixpress` most unique.

An aside: if you intend to expand on (1), e.g. [alternative store types](https://nix.dev/manual/nix/2.30/store/types/index.html), you might also consider writing a low-level Nix client to facilitate the implementation, kind of like `gert` for Git or `cmdstanr` for CmdStan. This might even help you maintain `rix`.

### Visualization

DAG visualizations greatly improve the user experience, but they are also a Pandora's Box of scope creep. `rixpress` already supports 3 backends for graphs: `visNetwork`, `ggdag`, and GraphViz (DOT; for CI). And each one is a magnet for feature requests.

To simplify the visualization feature set, what about using `mermaid.js` instead of `ggdag` or GraphViz? Mermaid graphs are just text, and they are very easy to generate without any additional R packages. For CI, you could use https://github.com/AlexanderGrooff/mermaid-ascii, which I think would produce graphs that are more readable and visually appealing than GraphViz can render (e.g. https://github.com/b-rodrigues/rixpress_demos/actions/runs/16252270236/job/45883546684#step:9:11).

`visNetwork` might only be necessary if you expect enormous pipelines whose graphs can only be explored interactively. If you do decide to keep `rixpress::rxp_visnetwork()`, I suggest keeping the feature set simple (which you have done already), and I suggest [disabling physics](https://datastorm-open.github.io/visNetwork/physic.html) to improve rendering performance for large graphs. Users may ask you for more features or aesthetic improvements, but I don't think an interactive `visNetwork` graph is the ideal place for nice-looking graphs. I suggest using `mermaid.js` for the pretty graphs.

### Workflow functions

From the examples, I see two patterns for setting up and running pipelines. In simple cases:

```r
list(
  rxp_r(...)
) |>
  rixpress()
```

but for Python projects such as [`r_py_xgboost`](https://github.com/b-rodrigues/rixpress_demos/tree/master/r_py_xgboost):

```r
list(
  rxp_py(...)
) |>
  rixpress(build = FALSE)

adjust_import(...)

rxp_make()
```

I think it would be clearer and more consistent to create a separate function (maybe `rxp_populate()`) which runs the equivalent of `rixpress(build = FALSE)`. Then, if `rixpress()` itself is still needed, it could serve as the equivalent of `rxp_populate()` + `rxp_make()`.

`add_import()` and `adjust_import()` feel a bit awkward as separate steps. You might instead consider an interface like:

```r
list(
  rxp_py(...)
) |>
rxp_populate(
  derivations,
  py_imports = c(
    numpy = "from numpy import array, loadtxt",
    xgboost = "from xgboost import XGBClassifier"
  )
)
```

### Names for functions and classes

I have minor suggestions to make the names of functions more internally consistent:

* `export_nix_archive()` => `rxp_export_nix_archive()`
* `import_nix_archive()` => `rxp_import_nix_archive()`
* `generate_dag()` => `rxp_generate_dag()` or `rxp_write_dag()` or `rxp_save_dag()`

(You may not need `dag_for_ci()`, `add_import()`, or `adjust_import()` if you agree with my suggestions from earlier.)

In addition, functions like `rxp_r()` produce an object of class `"derivation"`. I suggest renaming it to something like `"rxp_derivation"` so it does not conflict with e.g. mathematical packages with their own kinds of "derivations".

### Installation experience

I am new to Nix, `rix`, and `rixpress`, and I began by installing the toolchain from scratch on an M2 Macbook Pro with OS 15.5. This is my work computer, so it has more security restrictions than a regular personal computer.

I followed the [`rix` setup guide for macOS](https://docs.ropensci.org/rix/articles/b2-setting-up-and-using-rix-on-macos.html), which was clear and comprehensive. The [`curl` command](https://docs.ropensci.org/rix/articles/b2-setting-up-and-using-rix-on-macos.html#installing-nix) successfully downloaded the Determinate Systems installer, but the installer itself failed. First I realized I needed to run it with `sudo`, but even that failed. Nix installed successfully when I navigated a browser <https://docs.determinate.systems/>, manually downloaded the installer, and double-clicked it to run it. Maybe consider updating the [vignette](https://docs.ropensci.org/rix/articles/b2-setting-up-and-using-rix-on-macos.html#installing-nix) to mention that the point-and-click route is possible?

Afterwards, I installed `cachix` and ran `cachix use rstats-on-nix`. `library(rix)` initially showed [these warning messages](https://github.com/ropensci/rix/blob/4379be6d0cc17ac6b82969c3b30db1fb2877e20c/R/zzz.R#L24-L28), but `rix::setup_cachix()` silenced them. The next `library(rix)` gave me a warning about an incomplete final line in `~/.config/nix/nix.conf`, which I solved by manually opening the text file and adding a line break. I suggest ensuring `rix::setup_cachix()` leaves a terminating newline character in `~/.config/nix/nix.conf`.

### Storage

I really like the build logs feature you describe in https://b-rodrigues.github.io/rixpress/articles/g-logs.html. Over multiple pipeline runs, however, storage may accumulate, especially because Nix uses content-addressable storage (by hash). It may help to describe in that vignette how users can leverage the garbage collection features of Nix to clear out the data the no longer need.

### Multi-line expressions

The following pipeline succeeds:

```r
library(rixpress)
list(
  rxp_r(
    name = derivation,
    expr = 1 + 1
  )
) |>
  rixpress()
```

But a similar one fails:

```r
library(rixpress)
list(
  rxp_r(
    name = derivation,
    expr = {
      message("Running derivation")
      1 + 1
    }
  )
) |>
  rixpress()
```

with the error message:

```r
Error: unexpected numeric constant in "  derivation <- {     message('Running derivation')     1
```

Same if I add a semicolon after the `message()` statement. I expect this `rxp_r()` etc. do not support multi-line expressions. I would suggest either adding this support or requiring expressions to be pure function calls.

### Testing

I recommend including `skip_if_not_installed()` statements in tests where `Suggests:` packages are used (such as `mockery` and `reticulate`). In addition, when I ran the tests locally, one test threw a warning:

```r
Warning (test-generate_libraries_from_nix.R:42:3): generate_py_libraries_from_nix: generate Py script by parsing default.nix
Python packages have been requested, but 'reticulate' is not in your list of R packages. If you want to handle Python objects from your R session, consider adding 'reticulate' to the list of R packages.
Backtrace
    ▆
 1. └─rix::rix(...) at test-generate_libraries_from_nix.R:42:3
```

Adding `r_pkgs = reticulate` to `rix::rix()` in https://github.com/b-rodrigues/rixpress/blob/ea052f4a024bb47705bf186541380d5febac279e/tests/testthat/test-generate_libraries_from_nix.R#L42 should remove it.

Test coverage from `covr` is lower than I normally see in packages, but I really like your approach to offload to https://github.com/b-rodrigues/rixpress_demos. If the number of projects in that repo grows unmanageable at some point, you might consider creating a new GitHub org for them like https://github.com/nf-core does for Nextflow.

### Checks

When I ran `devtools::check()` locally, I saw: the note:

```
✔  checking for non-standard things in the check directory
N  checking for detritus in the temp directory
   Found the following files/directories:
     ‘RtmptcpOyn_repo_hash_url_jnlhe’
```

I have been flagged for this before when trying to submit packages to CRAN.

### Lints

On my local machine, `devtools::lint()` shows many lints, including:

```r
data-raw/gen_pipeline.R:6:3: style: [quotes_linter] Only use double-quotes.
  'mtcars.csv',
  ^~~~~~~~~~~~
data-raw/gen_pipeline.R:43:2: style: [commented_code_linter] Remove commented code.
#rxp_make()
 ^~~~~~~~~~
data-raw/jl_example/functions.R:1:34: style: [brace_linter] There should be a space before an opening curly brace.
prepare_data <- function(laplace){
                                 ^
```

I have never used the Air formatter, and I do see you have https://github.com/b-rodrigues/rixpress/blob/main/.github/workflows/style-with-air.yaml, so please disregard if there is an inherent conflict between Air and `lintr`.

`devtools::spell_check()` has many findings, including:

```
> devtools::spell_check()
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'.
  WORD                FOUND IN
’s                README.md:18
al                  a-intro-concepts.Rmd:124
Analysing           d-polyglot.Rmd:34
autoplay            b-core-functions.Rmd:272
buildInputs         make_derivation_snippet.Rd:20
cachix              d-polyglot.Rmd:46
Cachix              d-polyglot.Rmd:62
cancelled           rxp_init.Rd:17
ci                  dag_for_ci.Rd:36,40
                    generate_dag.Rd:29,33
                    rxp_ga.Rd:29,33
cmdstanr            f-cmdstanr.Rmd:2
configurePhase      make_derivation_snippet.Rd:20
cryptographic       a-intro-concepts.Rmd:164,171,214,230
CTRL                c-tutorial.Rmd:105
                    d-polyglot.Rmd:107
                    d2-polyglot-julia.Rmd:82
deriv               rixpress.Rd:18
...
```

You can exclude specific false positives in `inst/WORDLIST`.

On my machine, `urlchecker::url_check()` shows:

```
✖ Error: vignettes/a-intro-concepts.Rmd:32:21
403: Forbidden
spectrum/continuum](https://www.researchgate.net/figure/Reproducibility-spectrum-as-Peng-2011-stated_fig1_354765302),
                    ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
```

This could be because of the extra network security in my work environment, or it could be because ResearchGate has checks for bots. In the latter case, CRAN might flag the URL.

### Miscellaneous suggestions

* In `rxp_copy()`, I see `Sys.chmod(all_files, mode = "777")`, which could be risky on shared file systems. Is there a more restricted permission set that would still work?
* For `rxp_r_file()`, [the implicit `roxygen2` `@title` tag is `"rxp_r_file"`](https://github.com/b-rodrigues/rixpress/blob/ea052f4a024bb47705bf186541380d5febac279e/R/derivs.R#L681). I suggest a more descriptive name. Same for `rxp_py_file()`.
* Please consider changing the name of the default branch from `"master"` to `"main"` in https://github.com/b-rodrigues/rixpress_demos.
* `nix-store --realize` has [many options for `--verbose`](https://nix.dev/manual/nix/2.30/command-ref/nix-store/realise#opt-verbose). I suggest making the `verbose` argument of `rxp_make()` an integer to support this existing functionality. There are [many more features you could consider](https://docs.ropensci.org/targets/reference/index.html#progress) for helping users monitor pipelines, some of which are more feasible than others, and this one seems like the lowest-hanging fruit.
* In `rixpress_demos/r_multi_envs`, I recommend a more formal/safe choice for the meme image.
* Instead of prefixes to control the order vignettes are listed, you could consider relying on pkgdown yaml for this, e.g. https://github.com/wlandau/crew/blob/728c45536d58faf1794e2a16c469fdce4a815176/_pkgdown.yml#L6-L19.
