### 2025-08-16 b2d3fbd
first draft
