### 2025-08-18 6f85304
ordering of vignettes using pkgdown
