+++ vignettes/cmdstanr.Rmd
@@ -2,7 +2,7 @@
 title: "Using {cmdstanr} with {rixpress}"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{f-cmdstanr}
+  %\VignetteIndexEntry{cmdstanr}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
