+++ vignettes/polyglot.Rmd
@@ -15,7 +15,7 @@ knitr::opts_chunk$set(
 ```
 
 This vignette demonstrates how to build a polyglot pipeline and assumes you've
-read `vignette("b-core-functions")`.
+read `vignette("core-functions")`.
 
 For a video version of this vignette, [click here](https://youtu.be/LYtN1aOsTWQ).
 
