### 2025-08-21 c8938e7
Implement rxp_populate() function with inline Python import adjustments
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
