+++ README.Rmd
@@ -88,6 +88,34 @@ stage using:
 rxp_make()
 ```
 
+## Two-step workflow for Python projects
+
+For Python projects or when you need to adjust Python import statements, you can use the two-step workflow with `rxp_populate()`:
+
+```r
+# Define Python derivations
+derivs <- list(
+  rxp_py(data_load, "df = pd.read_csv('data.csv')"),
+  rxp_py(image_proc, "img = Image.open('photo.jpg')")
+)
+
+# Step 1: Generate pipeline with Python import adjustments
+rxp_populate(
+  derivs = derivs,
+  py_imports = list(
+    "import pandas" = "import pandas as pd",
+    "import pillow" = "from PIL import Image"
+  )
+)
+
+# Step 2: Build the pipeline
+rxp_make()
+```
+
+This approach is especially useful for packages like `pillow` (which requires `from PIL import Image` instead of `import pillow`) or when you want to use common aliases like `import numpy as np`.
+
+## Environment setup
+
 The build process assumes the presence of a `default.nix` file which defines the
 computational environment the pipeline runs in; this `default.nix` file can be
 generated by using the `{rix}` package. This `default.nix` defines an
