+++ NAMESPACE
@@ -17,6 +17,7 @@ export(rxp_jl)
 export(rxp_list_logs)
 export(rxp_load)
 export(rxp_make)
+export(rxp_populate)
 export(rxp_py)
 export(rxp_py2r)
 export(rxp_py_file)
