+++ tests/testthat/_snaps/rxp_populate/default_libraries.py
@@ -1,0 +1,2 @@
+import pickle
+from xgboost import XGBClassifier

[…]
