+++ man/rxp_populate.Rd
@@ -2,78 +2,49 @@
 % Please edit documentation in R/rxp_populate.R
 \name{rxp_populate}
 \alias{rxp_populate}
-\title{Generate Nix Pipeline Code without Building}
+\title{Populate the pipeline (generate files) without building}
 \usage{
-rxp_populate(derivs, project_path = ".", py_imports = NULL)
+rxp_populate(derivs, project_path = ".", py_imports = NULL, ...)
 }
 \arguments{
-\item{derivs}{A list of derivation objects, where each object is a list of
-five elements:
-\itemize{
-\item \verb{$name}, character, name of the derivation
-\item \verb{$snippet}, character, the nix code snippet to build this derivation
-\item \verb{$type}, character, can be R, Python or Quarto
-\item \verb{$additional_files}, character vector of paths to files to make
-available to build sandbox
-\item \verb{$nix_env}, character, path to Nix environment to build this derivation
-}
-A single deriv is the output of \code{rxp_r()}, \code{rxp_qmd()} or \code{rxp_py()}
-function.}
+\item{derivs}{A list of derivation objects created by rxp_*() helpers.}
 
-\item{project_path}{Path to root of project, defaults to ".".}
+\item{project_path}{Path to the project root. Defaults to ".".}
 
-\item{py_imports}{Named list for Python import adjustments. Names should be
-the old import statements to replace, and values should be the new import
-statements. For example: \code{list("import pillow" = "from PIL import Image", "import numpy" = "import numpy as np")}. If NULL (default), no import
-adjustments are made.}
+\item{py_imports}{Named character vector of Python import rewrites. Names are
+the base modules that rixpress auto-imports as "import \if{html}{\out{<name>}}", and values
+are the desired import lines. For example: c(numpy = "import numpy as np",
+pillow = "from PIL import Image"). Each entry is applied by replacing
+"import \if{html}{\out{<name>}}" with the provided string across generated _rixpress Python
+library files.}
+
+\item{...}{Ignored when build = FALSE; kept for signature parity with
+rixpress().}
 }
 \value{
-Nothing, writes a file called \code{pipeline.nix} with the Nix code to
-build the pipeline.
+Invisibly returns NULL. Side effects: writes generated files.
 }
 \description{
-This function generates a \code{pipeline.nix} file based on a list of derivation
-objects but does not build the pipeline. It is equivalent to calling
-\code{rixpress(build = FALSE)} but with additional support for inline Python
-import adjustments. After generating the pipeline, you can build it using
-\code{rxp_make()}.
-}
-\details{
-This function provides a two-step workflow that is particularly useful for
-Python projects:
-\enumerate{
-\item \code{rxp_populate()} - Generate the pipeline and optionally adjust Python imports
-\item \code{rxp_make()} - Build the pipeline
-}
-
-The \code{py_imports} parameter allows you to make common Python import adjustments
-inline, without needing to call \code{adjust_import()} separately. This is especially
-useful for packages like \code{pillow} (which needs \code{from PIL import Image} instead
-of \code{import pillow}) or when you want to use common aliases like \code{import numpy as np}.
+rxp_populate() generates the pipeline assets (e.g., _rixpress/, pipeline.nix,
+dag.json) but does not build. It is equivalent to calling rixpress(build =
+FALSE). Additionally, you can inline Python import adjustments via
+\code{py_imports}.
 }
 \examples{
 \dontrun{
[…]
