+++ pkgdown.yml
@@ -11,7 +11,7 @@ articles:
   polyglot: polyglot.html
   scope: scope.html
   tutorial: tutorial.html
-last_built: 2025-08-21T16:45Z
+last_built: 2025-08-21T18:54Z
 urls:
   reference: https://b-rodrigues.github.io/rixpress/reference
   article: https://b-rodrigues.github.io/rixpress/articles
