### 2025-08-21 283c968
Apply British English spelling fixes in R/ directory files
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
