### 2025-08-21 bd46e56
Fix British English spellings and typos in vignettes
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
