+++ R/derivs.R
@@ -130,8 +130,9 @@ rxp_r <- function(
     env_exports <- paste(
       vapply(
         names(env_var),
-        function(var_name)
-          sprintf("export %s=%s", var_name, env_var[[var_name]]),
+        function(var_name) {
+          sprintf("export %s=%s", var_name, env_var[[var_name]])
+        },
         character(1)
       ),
       collapse = "\n      "
@@ -270,8 +271,9 @@ rxp_py <- function(
       out_name
     )
   } else {
-    if (!is.character(serialize_function))
+    if (!is.character(serialize_function)) {
       stop("serialize_function must be a character string or NULL")
+    }
     serialize_str <- sprintf(
       "%s(globals()['%s'], '%s')",
       serialize_function,
@@ -284,8 +286,9 @@ rxp_py <- function(
   if (is.null(unserialize_function)) {
     unserialize_str <- "pickle.load"
   } else {
-    if (!is.character(unserialize_function))
+    if (!is.character(unserialize_function)) {
       stop("unserialize_function must be a character string or NULL")
+    }
     unserialize_str <- unserialize_function
   }
 
@@ -425,7 +428,9 @@ rxp_qmd <- function(
     matches <- gregexpr(pattern, content_str)
     full_matches <- regmatches(content_str, matches)[[1]]
 
-    if (length(full_matches) == 0) return(data.frame())
+    if (length(full_matches) == 0) {
+      return(data.frame())
+    }
 
     results <- data.frame(
       full_match = character(0),
@@ -541,8 +546,9 @@ rxp_qmd <- function(
     env_exports <- paste(
       vapply(
         names(env_var),
-        function(var_name)
-          sprintf("      export %s=%s", var_name, env_var[[var_name]]),
+        function(var_name) {
+          sprintf("      export %s=%s", var_name, env_var[[var_name]])
+        },
         character(1)
       ),
       collapse = "\n"
@@ -557,8 +563,11 @@ rxp_qmd <- function(
     "      export HOME=$PWD/home",
     "      export RETICULATE_PYTHON=${defaultPkgs.python3}/bin/python",
     env_exports,
-    if (length(sub_cmds) > 0)
-      paste("      ", sub_cmds, sep = "", collapse = "\n") else "",
+    if (length(sub_cmds) > 0) {
+      paste("      ", sub_cmds, sep = "", collapse = "\n")
+    } else {
+      ""
+    },
     sprintf("      quarto render %s %s --output-dir $out", qmd_file, args),
     sep = "\n"
   )
@@ -667,8 +676,9 @@ rxp_file_common <- function(
     env_exports <- paste(
       vapply(
         names(env_var),
-        function(var_name)
-          sprintf("export %s=%s", var_name, env_var[[var_name]]),
+        function(var_name) {
+          sprintf("export %s=%s", var_name, env_var[[var_name]])
+        },
         character(1)
       ),
       collapse = "\n      "
@@ -703,8 +713,6 @@ rxp_file_common <- function(
     structure(class = "derivation")
 }
 
-#' rxp_r_file
-#'
 #' Creates a Nix expression that reads in a file (or folder of data) using R.
 #'
 #' @family derivations
@@ -806,8 +814,6 @@ rxp_r_file <- function(
   )
 }
 
-#' rxp_py_file
-#'
 #' Creates a Nix expression that reads in a file (or folder of data) using
 #' Python.
 #'
@@ -1122,8 +1128,9 @@ rxp_rmd <- function(
     env_exports <- paste(
       vapply(
         names(env_var),
-        function(var_name)
-          sprintf("      export %s=%s", var_name, env_var[[var_name]]),
[…]
