+++ DESCRIPTION
@@ -22,6 +22,7 @@ Imports:
     jsonlite,
     processx
 RoxygenNote: 7.3.2
+Language: en-GB
 Suggests:
     dplyr,
     ggdag,
