### 2025-08-24 472bb5e
better doc for rxp_inspect
