+++ man/export_nix_archive.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rxp_make.R
-\name{export_nix_archive}
-\alias{export_nix_archive}
+\name{rxp_export_artifacts}
+\alias{rxp_export_artifacts}
 \title{Export Nix store paths to an archive}
 \usage{
-export_nix_archive(
+rxp_export_artifacts(
   archive_file = "_rixpress/pipeline_outputs.nar",
   which_log = NULL,
   project_path = "."
@@ -32,10 +32,10 @@ its Nix store.
 \examples{
 \dontrun{
   # Export the most recent build to the default location
-  export_nix_archive()
+  rxp_export_artifacts()
 
   # Export a specific build to a custom location
-  export_nix_archive(
+  rxp_export_artifacts(
     archive_file = "my_archive.nar",
     which_log = "20250510"
   )
@@ -43,6 +43,6 @@ its Nix store.
 }
 \seealso{
 Other archive caching functions: 
-\code{\link{import_nix_archive}()}
+\code{\link{rxp_import_artifacts}()}
 }
 \concept{archive caching functions}
