+++ README.Rmd
@@ -159,7 +159,7 @@ read it from there.
 
 You can export the cache into a file and easily import it on another machine (or
 on CI) to avoid having to rebuild everything from scratch using
-`export_nix_archive()` and `import_nix_archive()` respectively.
+`rxp_export_artifacts()` and `rxp_import_artifacts()` respectively.
 
 `rixpress()` is very flexible; please consult [this
 repository](https://github.com/b-rodrigues/rixpress_demos/tree/master) which
