+++ man/import_nix_archive.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rxp_make.R
-\name{import_nix_archive}
-\alias{import_nix_archive}
+\name{rxp_import_artifacts}
+\alias{rxp_import_artifacts}
 \title{Import Nix store paths from an archive}
 \usage{
-import_nix_archive(archive_file = "_rixpress/pipeline_outputs.nar")
+rxp_import_artifacts(archive_file = "_rixpress/pipeline_outputs.nar")
 }
 \arguments{
 \item{archive_file}{Character, path to the archive, defaults to
@@ -20,14 +20,14 @@ Useful for transferring built outputs between machines.
 \examples{
 \dontrun{
   # Import from the default archive location
-  import_nix_archive()
+  rxp_import_artifacts()
 
   # Import from a custom archive file
-  import_nix_archive("path/to/my_archive.nar")
+  rxp_import_artifacts("path/to/my_archive.nar")
 }
 }
 \seealso{
 Other archive caching functions: 
-\code{\link{export_nix_archive}()}
+\code{\link{rxp_export_artifacts}()}
 }
 \concept{archive caching functions}
