+++ NAMESPACE
@@ -4,9 +4,9 @@ S3method(print,derivation)
 export(add_import)
 export(adjust_import)
 export(dag_for_ci)
-export(export_nix_archive)
+export(rxp_export_artifacts)
 export(generate_dag)
-export(import_nix_archive)
+export(rxp_import_artifacts)
 export(rixpress)
 export(rxp_copy)
 export(rxp_ga)
