+++ NAMESPACE
@@ -4,13 +4,13 @@ S3method(print,derivation)
 export(add_import)
 export(adjust_import)
 export(dag_for_ci)
-export(rxp_export_artifacts)
 export(generate_dag)
-export(rxp_import_artifacts)
 export(rixpress)
 export(rxp_copy)
+export(rxp_export_artifacts)
 export(rxp_ga)
 export(rxp_ggdag)
+export(rxp_import_artifacts)
 export(rxp_init)
 export(rxp_inspect)
 export(rxp_jl)
