### 2025-08-24 04fd8e7
generate_dag -> rxp_write_dag
