+++ man/rxp_write_dag.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/generate_dag.R
-\name{generate_dag}
-\alias{generate_dag}
+\name{rxp_write_dag}
+\alias{rxp_write_dag}
 \title{Generate a DAG from a list of derivations}
 \usage{
-generate_dag(rxp_list, output_file = "_rixpress/dag.json")
+rxp_write_dag(rxp_list, output_file = "_rixpress/dag.json")
 }
 \arguments{
 \item{rxp_list}{A list of derivations.}
@@ -22,7 +22,7 @@ by \code{rixpress()}.
 }
 \examples{
 \dontrun{
-  generate_dag(rxp_list)
+  rxp_write_dag(rxp_list)
 }
 }
 \seealso{
