+++ NAMESPACE
@@ -4,7 +4,6 @@ S3method(print,derivation)
 export(add_import)
 export(adjust_import)
 export(dag_for_ci)
-export(generate_dag)
 export(rixpress)
 export(rxp_copy)
 export(rxp_export_artifacts)
@@ -28,6 +27,7 @@ export(rxp_r_file)
 export(rxp_read)
 export(rxp_rmd)
 export(rxp_visnetwork)
+export(rxp_write_dag)
 importFrom(igraph,V)
 importFrom(igraph,delete_vertex_attr)
 importFrom(igraph,graph_from_data_frame)
