+++ R/rxp_ga.R
@@ -40,7 +40,7 @@ rxp_ga <- function() {
   file.copy(source, path, overwrite = TRUE)
 
   # Generate dag.dot to view in CI
-  dag_for_ci()
+  rxp_dag_for_ci()
 
   if (identical(Sys.getenv("TESTTHAT"), "false")) {
     message("GitHub Actions workflow file saved to: ", path)
