+++ man/rxp_ga.Rd
@@ -27,7 +27,7 @@ to give read and write permissions to the GitHub Actions bot.
 }
 \seealso{
 Other ci utilities: 
-\code{\link{dag_for_ci}()},
+\code{\link{rxp_dag_for_ci}()},
 \code{\link{rxp_write_dag}()}
 }
 \concept{ci utilities}
