+++ man/rxp_dag_for_ci.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plot_dag.R
-\name{dag_for_ci}
-\alias{dag_for_ci}
+\name{rxp_dag_for_ci}
+\alias{rxp_dag_for_ci}
 \title{Export DAG of pipeline and prepare it for rendering on CI}
 \usage{
-dag_for_ci(
+rxp_dag_for_ci(
   nodes_and_edges = get_nodes_edges(),
   output_file = "_rixpress/dag.dot"
 )
@@ -28,7 +28,7 @@ is called by \code{rxp_ga()}.
 \examples{
 \dontrun{
   # Generate the default _rixpress/dag.dot
-  dag_for_ci()
+  rxp_dag_for_ci()
 
 }
 }
