+++ NAMESPACE
@@ -3,9 +3,9 @@
 S3method(print,derivation)
 export(add_import)
 export(adjust_import)
-export(dag_for_ci)
 export(rixpress)
 export(rxp_copy)
+export(rxp_dag_for_ci)
 export(rxp_export_artifacts)
 export(rxp_ga)
 export(rxp_ggdag)
