+++ R/plot_dag.R
@@ -53,7 +53,7 @@ rxp_dag_for_ci <- function(
 get_nodes_edges <- function(path_dag = "_rixpress/dag.json") {
   # Check if the DAG file exists
   if (!file.exists(path_dag)) {
-    stop("dag.json missing! Did you run 'rixpress()'?")
+    stop("dag.json missing! Did you run 'rxp_populate()'?")
   }
 
   json_data <- fromJSON(path_dag)
