+++ data-raw/py2r_example/gen-pipeline.R
@@ -56,7 +56,7 @@ doc <- rxp_qmd(
 
 rxp_list <- list(d0, d1, d2, d3, d3_1, d4, d4_1, d5, doc)
 
-rixpress(rxp_list, project_path = ".", build = FALSE)
+rxp_populate(rxp_list, project_path = ".", build = FALSE)
 
 rxp_make()
 
