+++ man/rxp_populate.Rd
@@ -1,8 +1,8 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rixpress.R
+% Please edit documentation in R/rxp_populate.R
 \name{rxp_populate}
 \alias{rxp_populate}
-\title{Generate Nix Pipeline Code with a Generic Default Target}
+\title{Generate Nix Pipeline Code}
 \usage{
 rxp_populate(derivs, project_path = ".", build = FALSE, py_imports = NULL, ...)
 }
