### 2025-08-25 24b9647
rewritings
