### 2025-08-25 f5c670c
derivation -> rxp_derivation
