+++ tests/testthat/test-derive_r.R
@@ -8,12 +8,13 @@ test_that("rxp_r: generates correct list", {
         "snippet" = '  mtcars_am = makeRDerivation {\n    name = \"mtcars_am\";\n    buildInputs = defaultBuildInputs;\n    configurePhase = defaultConfigurePhase;\n    buildPhase = \'\'\n      Rscript -e \"\n        source(\'libraries.R\')\n        mtcars_am <- dplyr::filter(mtcars, am == 1)\n        saveRDS(mtcars_am, \'mtcars_am\')\"\n    \'\';\n  };',
         "type" = "rxp_r",
         "additional_files" = "",
+        "user_functions" = "",
         "nix_env" = "default.nix",
         "serialize_function" = "saveRDS",
         "unserialize_function" = "readRDS",
         "env_var" = NULL
       ),
-      class = "derivation"
+      class = "rxp_derivation"
     )
   )
 })
@@ -28,12 +29,13 @@ test_that("rxp_py: generates correct list", {
         "snippet" = '  mtcars_pl_am = makePyDerivation {\n    name = \"mtcars_pl_am\";\n    buildInputs = defaultBuildInputs;\n    configurePhase = defaultConfigurePhase;\n    buildPhase = \'\'\n      python -c "\nexec(open(\'libraries.py\').read())\nexec(\'mtcars_pl_am = mtcars_pl.filter(pl.col(\\\'am\\\') == 1)\')\nwith open(\'mtcars_pl_am\', \'wb\') as f: pickle.dump(globals()[\'mtcars_pl_am\'], f)\n"\n    \'\';\n  };',
         "type" = "rxp_py",
         "additional_files" = "",
+        "user_functions" = "",
         "nix_env" = "default.nix",
         "serialize_function" = "with open('mtcars_pl_am', 'wb') as f: pickle.dump(globals()['mtcars_pl_am'], f)",
         "unserialize_function" = "pickle.load",
         "env_var" = NULL
       ),
-      class = "derivation"
+      class = "rxp_derivation"
     )
   )
 })
@@ -55,12 +57,13 @@ test_that("rxp_py: custom serialization functions work", {
         "snippet" = '  mtcars_pl_am = makePyDerivation {\n    name = \"mtcars_pl_am\";\n    buildInputs = defaultBuildInputs;\n    configurePhase = defaultConfigurePhase;\n    buildPhase = \'\'\n      python -c "\nexec(open(\'libraries.py\').read())\nexec(\'mtcars_pl_am = mtcars_pl.filter(pl.col(\\\'am\\\') == 1)\')\ncustom_save(globals()[\'mtcars_pl_am\'], \'mtcars_pl_am\')\n"\n    \'\';\n  };',
         "type" = "rxp_py",
         "additional_files" = "",
+        "user_functions" = "",
         "nix_env" = "default.nix",
         "serialize_function" = "custom_save(globals()['mtcars_pl_am'], 'mtcars_pl_am')",
         "unserialize_function" = "custom_load",
         "env_var" = NULL
       ),
-      class = "derivation"
+      class = "rxp_derivation"
     )
   )
 })
@@ -101,7 +104,7 @@ test_that("rxp_qmd: generates correct list", {
         "args" = "--to pdf",
         "env_var" = NULL
       ),
-      class = "derivation"
+      class = "rxp_derivation"
     )
   )
 
@@ -125,7 +128,7 @@ test_that("rxp_r_file: single file reading works", {
     "nix_env" = d1$nix_env,
     "copy_data_folder" = FALSE
   )
-  class(d1_subset) <- "derivation"
+  class(d1_subset) <- "rxp_derivation"
 
   testthat::expect_equal(
     d1_subset,
[…]
