+++ R/derivs.R
@@ -49,9 +49,14 @@ make_derivation_snippet <- function(
 #' @family derivations
 #' @param name Symbol, name of the derivation.
 #' @param expr R code to generate the expression.
-#' @param additional_files Character vector, additional files to include. Custom
-#'   functions must go into a script called "functions.R", and additional files
-#'   that need to be accessible during the build process can be named anything.
+#' @param additional_files Character vector, additional files to include
+#'   during the build process. For example, if a function expects a certain
+#'   file to be available, this is where you should include it.
+#' @param user_functions Character vector, user-defined functions to include.
+#'   This should be a script (or scripts) containing user-defined functions
+#'   to include during the build process for this derivation. It is recommended
+#'   to use one script per function, and only include the required script(s) in
+#'   the derivation.
 #' @param nix_env Character, path to the Nix environment file, default is
 #'   "default.nix".
 #' @param serialize_function Function, defaults to NULL. A function used to
@@ -102,11 +107,11 @@ rxp_r <- function(
   name,
   expr,
   additional_files = "",
+  user_functions = "",
   nix_env = "default.nix",
   serialize_function = NULL,
   unserialize_function = NULL,
-  env_var = NULL,
-  user_functions = NULL
+  env_var = NULL
 ) {
   out_name <- deparse1(substitute(name))
   expr_str <- deparse1(substitute(expr))
@@ -236,13 +241,13 @@ rxp_r <- function(
     snippet = snippet,
     type = "rxp_r",
     additional_files = additional_files,
+    user_functions = user_functions,
     nix_env = nix_env,
     serialize_function = serialize_str,
     unserialize_function = unserialize_str,
-    env_var = env_var,
-    user_functions = user_functions
+    env_var = env_var
   ) |>
-    structure(class = "derivation")
+    structure(class = "rxp_derivation")
 }
 
 #' Create a Nix expression running a Python function
@@ -250,9 +255,14 @@ rxp_r <- function(
 #' @family derivations
 #' @param name Symbol, name of the derivation.
 #' @param py_expr Character, Python code to generate the expression.
-#' @param additional_files Character vector, additional files to include. Custom
-#'   functions must go into a script called "functions.py", and additional files
-#'   that need to be accessible during the build process can be named anything.
+#' @param additional_files Character vector, additional files to include
+#'   during the build process. For example, if a function expects a certain
+#'   file to be available, this is where you should include it.
+#' @param user_functions Character vector, user-defined functions to include.
+#'   This should be a script (or scripts) containing user-defined functions
+#'   to include during the build process for this derivation. It is recommended
+#'   to use one script per function, and only include the required script(s) in
+#'   the derivation.
 #' @param nix_env Character, path to the Nix environment file, default is
 #'   "default.nix".
 #' @param serialize_function Character, defaults to NULL. The name of the Python
@@ -295,11 +305,11 @@ rxp_py <- function(
   name,
   py_expr,
   additional_files = "",
+  user_functions = "",
   nix_env = "default.nix",
   serialize_function = NULL,
   unserialize_function = NULL,
-  env_var = NULL,
-  user_functions = NULL # <-- NEW ARGUMENT
+  env_var = NULL
 ) {
   out_name <- deparse1(substitute(name))
   py_expr <- gsub("'", "\\'", py_expr, fixed = TRUE)
@@ -446,13 +456,13 @@ rxp_py <- function(
     snippet = snippet,
     type = "rxp_py",
     additional_files = additional_files,
[…]
