+++ man/rxp_copy.Rd
@@ -46,7 +46,7 @@ input, as is the case for \code{html} output.
 }
 \seealso{
 Other utilities: 
-\code{\link{print.derivation}()},
+\code{\link{print.rxp_derivation}()},
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
