### 2025-08-25 6b06f4c
better handling of dependencies
