+++ R/rxp_populate.R
@@ -402,15 +402,12 @@ in
 #' @noRd
 gen_pipeline <- function(dag_file, flat_pipeline) {
   dag <- jsonlite::read_json(dag_file)
-  pipeline <- flat_pipeline
+  pipeline_str <- paste(flat_pipeline, collapse = "\n")
 
   for (d in dag$derivations) {
     if (
       length(d$depends) == 0 ||
-        d$type == "rxp_qmd" ||
-        d$type == "rxp_rmd" ||
-        d$type == "rxp_py2r" ||
-        d$type == "rxp_r2py"
+        d$type %in% c("rxp_qmd", "rxp_rmd", "rxp_py2r", "rxp_r2py")
     ) {
       next
     }
@@ -420,121 +417,49 @@ gen_pipeline <- function(dag_file, flat_pipeline) {
     type <- d$type[1]
     unserialize_function <- d$unserialize_function
 
-    # Set parameters based on derivation type
+    # Define language-specific helpers
     if (type == "rxp_r") {
-      maker <- "makeRDerivation"
-      script_cmd <- "Rscript -e \""
-      load_line <- function(dep, indent, unserialize_function) {
-        paste0(
-          indent,
-          dep,
-          " <- ",
-          unserialize_function,
-          "('${",
-          dep,
-          "}/",
-          dep,
-          "')"
-        )
-      }
-    } else if (type == "rxp_py") {
-      maker <- "makePyDerivation"
-      script_cmd <- "python -c \""
-      load_line <- function(dep, indent, unserialize_function) {
-        path <- paste0("${", dep, "}/", dep)
-        if (unserialize_function == "pickle.load") {
-          paste0(
-            "with open('",
-            path,
-            "', 'rb') as f: ",
-            dep,
-            " = pickle.load(f)"
-          )
-        } else {
-          paste0(
-            dep,
-            " = ",
-            unserialize_function,
-            "('",
-            path,
-            "')"
-          )
-        }
-      }
-    } else if (type == "rxp_jl") {
-      maker <- "makeJlDerivation"
-      script_cmd <- "julia -e \""
-      load_line <- function(dep, indent, unserialize_function) {
-        path <- paste0("${", dep, "}/", dep)
-        paste0(
-          dep,
-          " = ",
-          unserialize_function,
-          "(\\\"",
-          path,
-          "\\\")"
-        )
-      }
+      # Add 'else if' blocks here for Python and Julia if needed
+      placeholder <- "# RIXPRESS_LOAD_DEPENDENCIES_HERE"
+      load_line_template <- "%s <- %s('${%s}/%s')" # obj <- readRDS('${obj}/obj')
+
+      load_lines <- vapply(
+        deps,
+        function(dep) {
+          sprintf(load_line_template, dep, unserialize_function, dep, dep)
+        },
+        character(1)
+      )
+
+      replacement_str <- paste(load_lines, collapse = "\n")
     } else {
-      warning("Unsupported type for derivation ", deriv_name)
-      next
+      next # Skip unsupported types for now
     }
 
-    # Locate the derivation block
-    pattern <- paste0("^\\s*", deriv_name, " = ", maker, " \\{")
-    start_idx <- grep(pattern, pipeline)
-    if (!length(start_idx)) {
-      warning("Derivation ", deriv_name, " not found")
-      next
-    }
-    start_idx <- start_idx[1]
-
-    # Find the end of the block
-    end_candidates <- grep("^\\s*};", pipeline)
-    block_end_idx <- end_candidates[end_candidates > start_idx][1]
-    if (is.na(block_end_idx)) {
-      warning("Block end for ", deriv_name, " not found")
-      next
-    }
-
[…]
