+++ R/derivs.R
@@ -416,6 +416,7 @@ rxp_py <- function(
     user_functions_copy_cmd,
     "python -c \"\n",
     "exec(open('libraries.py').read())\n",
+    "# RIXPRESS_PY_LOAD_DEPENDENCIES_HERE\n",
     user_import_cmd,
     "exec('",
     out_name,
@@ -1524,6 +1525,7 @@ rxp_jl <- function(
     user_functions_copy_cmd,
     "julia -e \"\n",
     "if isfile(\\\"libraries.jl\\\"); include(\\\"libraries.jl\\\"); end; ",
+    "# RIXPRESS_JL_LOAD_DEPENDENCIES_HERE; ",
     user_include_cmd,
     out_name,
     " = ",
