+++ data-raw/py_example/gen_pipeline.R
@@ -9,7 +9,7 @@ d1 <- rxp_py(
 d2 <- rxp_py(
   diabetes_head,
   'my_head(diabetes)',
-  additional_files = "functions.py"
+  user_functions = "functions.py"
 )
 d3 <- rxp_py(diabetes_tail, 'diabetes.tail()')
 d4 <- rxp_py(
@@ -19,17 +19,4 @@ d4 <- rxp_py(
 
 rxp_list <- list(d0, d1, d2, d3, d4)
 
-rxp_populate(rxp_list, project_path = ".")
-
-plot_dag()
-
-rxp_make()
-
-dag_obj <- plot_dag(return_igraph = TRUE)
-
-dag_obj <- set_vertex_attr(dag_obj, "label", value = V(dag_obj)$name)
-
-# Step 2: Delete the "name" attribute
-dag_obj <- delete_vertex_attr(dag_obj, "name")
-
-igraph::write_graph(dag_obj, file = "dag.dot", format = "dot")
+rxp_populate(rxp_list, project_path = ".", build = TRUE)
