### 2025-08-26 29d0947
handling of io and user_functions
