+++ R/rxp_io.R
@@ -1,0 +1,461 @@
+#' Creates a Nix expression that reads in a file (or folder of data) using R.
+#'
+#' @family derivations
+#' @param name Symbol, the name of the derivation.
+#' @param path Character, the file path to include (e.g., "data/mtcars.shp") or
+#'   a folder path (e.g., "data"). See details.
+#' @param read_function Function, an R function to read the data, taking one
+#'   argument (the path).
+#' @param user_functions Character vector, user-defined functions to include.
+#'   This should be a script (or scripts) containing user-defined functions
+#'   to include during the build process for this derivation. It is recommended
+#'   to use one script per function, and only include the required script(s) in
+#'   the derivation.
+#' @param nix_env Character, path to the Nix environment file, default is
+#'   "default.nix".
+#' @param copy_data_folder Logical, if TRUE then the entire folder is copied
+#'   recursively into the build sandbox.
+#' @param env_var List, defaults to NULL. A named list of environment variables
+#'   to set before running the R script, e.g., c(VAR = "hello"). Each entry will
+#'   be added as an export statement in the build phase.
+#' @details There are three ways to read in data in a rixpress pipeline: the
+#'   first is to point directly to a file, for example, `rxp_r_file(mtcars, path
+#'   = "data/mtcars.csv", read_function = read.csv)`. The second way is to point
+#'   to a file but to also include the files in the "data/" folder (the folder
+#'   can be named something else). This is needed when data is split between
+#'   several files, such as a shapefile which typically also needs other files
+#'   such as `.shx` and `.dbf` files. For this, `copy_data_folder` must be set
+#'   to `TRUE`. The last way to read in data, is to only point to a folder, and
+#'   use a function that recursively reads in all data. For example
+#'   `rxp_r_file(many_csvs, path = "data", read_function = \(x)(readr::read_csv(
+#'   list.files(x, full.names = TRUE, pattern = ".csv$"))))` the provided
+#'   anonymous function will read all the `.csv` files in the `data/` folder.
+#' @return An object of class derivation which inherits from lists.
+#' @examples
+#' \dontrun{
+#'   # Read a CSV file
+#'   rxp_r_file(
+#'     name = mtcars,
+#'     path = "data/mtcars.csv",
+#'     read_function = \(x) (read.csv(file = x, sep = "|"))
+#'   )
+#'
+#'   # Read all CSV files in a directory using a lambda function
+#'  rxp_r_file(
+#'    name = mtcars_r,
+#'    path = "data",
+#'    read_function = \(x)
+#'      (readr::read_delim(list.files(x, full.names = TRUE), delim = "|")),
+#'    copy_data_folder = TRUE
+#'  )
+#' }
+#' @export
+rxp_r_file <- function(
+  name,
+  path,
+  read_function,
+  user_functions = "",
+  nix_env = "default.nix",
+  copy_data_folder = FALSE,
+  env_var = NULL
+) {
+  out_name <- deparse1(substitute(name))
+  read_func_str <- deparse1(substitute(read_function))
+  read_func_str <- gsub("\"", "'", read_func_str)
+
+  # build copy command for user_functions
+  user_functions_copy_cmd <- ""
+  if (!is.null(user_functions) && length(user_functions) > 0) {
+    user_functions_clean <- user_functions[nzchar(user_functions)]
+    if (length(user_functions_clean) > 0) {
+      user_copy_lines <- vapply(
+        user_functions_clean,
[…]
