+++ R/derivs.R
@@ -466,6 +466,239 @@ rxp_py <- function(
     structure(class = "rxp_derivation")
 }
 
+
+#' Create a Nix expression running a Julia function
+#'
+#' @param name Symbol, name of the derivation.
+#' @param jl_expr Character, Julia code to generate the expression.
+#' @param additional_files Character vector, additional files to include
+#'   during the build process. For example, if a function expects a certain
+#'   file to be available, this is where you should include it.
+#' @param user_functions Character vector, user-defined functions to include.
+#'   This should be a script (or scripts) containing user-defined functions
+#'   to include during the build process for this derivation. It is recommended
+#'   to use one script per function, and only include the required script(s) in
+#'   the derivation.
+#' @param nix_env Character, path to the Nix environment file, default is
+#'   "default.nix".
+#' @param serialize_function Character, defaults to NULL. The name of the Julia
+#'   function used to serialize the object. It must accept two arguments: the
+#'   object to serialize (first), and the target file path (second). If NULL,
+#'   the default behaviour uses the built‐in `Serialization.serialize` API. Define
+#'   any custom serializer in `functions.jl`.
+#' @param unserialize_function Character, defaults to NULL. The name of the Julia
+#'   function used to unserialize the object. It must accept one argument: the
+#'   file path. If NULL, the default is assumed to be `Serialization.deserialize`.
+#' @param env_var Character vector, defaults to NULL. A named vector of
+#'   environment variables to set before running the Julia script, e.g.,
+#'   `c("JULIA_DEPOT_PATH" = "/path/to/depot")`. Each entry will be added as
+#'   an `export` statement in the build phase.
+#' @details At a basic level,
+#'   `rxp_jl(filtered_data, "filter(df, :col .> 10)")` is equivalent to
+#'   `filtered_data = filter(df, :col .> 10)` in Julia. `rxp_jl()` generates the
+#'   required Nix boilerplate to output a so‐called "derivation" in Nix jargon.
+#'   A Nix derivation is a recipe that defines how to create an output (in this
+#'   case `filtered_data`) including its dependencies, build steps, and output
+#'   paths.
+#' @return An object of class derivation which inherits from lists.
+#' @examples
+#' \dontrun{
+#' # Basic usage, no custom serializer
+#' rxp_jl(
+#'   name = filtered_df,
+#'   jl_expr = "filter(df, :col .> 10)"
+#' )
+#'
+#' # Custom serialization: assume `save_my_obj(obj, path)` is defined in functions.jl
+#' rxp_jl(
+#'   name = model_output,
+#'   jl_expr = "train_model(data)",
+#'   serialize_function = "save_my_obj",
+#'   additional_files = "functions.jl"
+#' )
+#' }
+#' @family derivations
+#' @export
+rxp_jl <- function(
+  name,
+  jl_expr,
+  additional_files = "",
+  user_functions = "",
+  nix_env = "default.nix",
+  serialize_function = NULL,
+  unserialize_function = NULL,
+  env_var = NULL
+) {
+  out_name <- deparse1(substitute(name))
+  # Escape double quotes for Julia one-liner
+  jl_expr_escaped <- gsub("\"", "\\\\\"", jl_expr)
+
+  # Determine which serialize function to call
+  if (is.null(serialize_function)) {
+    # Default: use built-in Serialization.serialize
+    serialize_str <- paste0(
+      "using Serialization; ",
+      "io = open(\\\"",
+      out_name,
+      "\\\", \\\"w\\\"); ",
[…]
