### 2025-08-26 9026083
refactor rxp_r_file and rxp_py_file
