+++ R/rxp_io.R
@@ -1,422 +1,273 @@
+#' Clean user functions vector
+#'
+#' Removes empty character entries from the provided `user_functions` vector.
+#'
+#' @param user_functions Character vector of user-defined function script paths.
+#' @return Character vector without empty strings.
+clean_user_functions <- function(user_functions) {
+  if (is.null(user_functions)) return(character(0))
+  user_functions[nzchar(user_functions)]
+}
+
+#' Build copy commands for user functions
+#'
+#' Generates shell commands to copy user-defined function scripts into the
+#' build sandbox.
+#'
+#' @param user_functions Character vector of script file names.
+#' @return A string of `cp` commands, suitable for embedding in a build phase.
+build_user_copy_cmd <- function(user_functions) {
+  if (length(user_functions) == 0) return("")
+  paste0(
+    paste(
+      sprintf("cp ${./%s} %s", user_functions, user_functions),
+      collapse = "\n      "
+    ),
+    "\n      "
+  )
+}
+
+#' Build code import/source commands for user functions
+#'
+#' Creates language-specific commands to load user-defined function scripts.
+#'
+#' @param user_functions Character vector of script file names.
+#' @param lang Language string, `"R"` or `"Python"`.
+#' @return A string of import/source statements.
+build_user_code_cmd <- function(user_functions, lang) {
+  if (length(user_functions) == 0) return("")
+
+  fmt <- switch(
+    lang,
+    R      = "source('%s')",
+    Py = "exec(open('%s').read())"
+  )
+
+  paste0(
+    paste(sprintf(fmt, user_functions), collapse = "\n        "),
+    "\n        "
+  )
+}
+
+#' Build environment variable export commands
+#'
+#' Generates shell export commands for setting environment variables before
+#' running the build script.
+#'
+#' @param env_var Named list of environment variables.
+#' @return A string of `export` commands with line breaks, or `""` if none.
+build_env_exports <- function(env_var) {
+  if (is.null(env_var) || length(env_var) == 0) return("")
+  lines <- vapply(names(env_var), function(var) {
+    sprintf("export %s=%s", var, env_var[[var]])
+  }, character(1))
+  paste0(paste(lines, collapse = "\n      "), "\n      ")
+}
+
+#' Build Nix src part
+#'
+#' Creates the `src` attribute for the Nix derivation. Handles both remote URLs
+#' (using `nix-prefetch-url`) and local paths (with optional user function
+#' scripts).
+#'
+#' @param actual_path Character path to file or directory.
+#' @param user_functions Character vector of additional files to include.
+#' @return A string with the `src` Nix expression.
+build_src_part <- function(actual_path, user_functions) {
+  if (grepl("^https?://", actual_path)) {
+    hash <- tryCatch(
+      system(paste("nix-prefetch-url", shQuote(actual_path)), intern = TRUE),
+      error = function(e) stop("Failed to run nix-prefetch-url: ", e$message)
+    )
+    if (length(hash) == 0 || hash == "")
+      stop("nix-prefetch-url did not return a hash for URL: ", actual_path)
+
+    sprintf(
+      "defaultPkgs.fetchurl {\n      url = \"%s\";\n      sha256 = \"%s\";\n    }",
+      actual_path,
+      trimws(hash[1])
+    )
+  } else if (length(user_functions) > 0) {
+    fileset_parts <- c(actual_path, user_functions)
+    fileset_nix <- paste0("./", fileset_parts, collapse = " ")
+    sprintf(
+      "defaultPkgs.lib.fileset.toSource {\n      root = ./.;\n      fileset = defaultPkgs.lib.fileset.unions [ %s ];\n    }",
+      fileset_nix
+    )
[…]
