### 2025-08-26 9bff40c
Fix single file copy bug in build_phase function
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
