### 2025-08-26 8b0b86f
Fix single file copy command when user_functions are present
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
