### 2025-08-27 2d79d1a
Fix gen_pipeline to correctly handle derivations with no dependencies
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
