### 2025-08-27 ad05189
Add comprehensive test for dependency loading fix
Co-authored-by: b-rodrigues <2998834+b-rodrigues@users.noreply.github.com>
