+++ tests/testthat/test-rixpress.R
@@ -216,3 +216,103 @@ test_that("rixpress correctly handles dependencies", {
     name = "pipeline_with_dependencies.nix"
   )
 })
+
+test_that("gen_pipeline correctly handles derivations with no dependencies", {
+  # Create a test DAG with derivations having no dependencies and some with dependencies
+  dag_content <- list(
+    derivations = list(
+      list(
+        deriv_name = "no_deps_r",
+        depends = list(), # No dependencies - key test case
+        unserialize_function = "readRDS",
+        type = "rxp_r"
+      ),
+      list(
+        deriv_name = "with_deps_r", 
+        depends = list("no_deps_r"), # Depends on no_deps_r
+        unserialize_function = "readRDS",
+        type = "rxp_r"
+      ),
+      list(
+        deriv_name = "no_deps_py",
+        depends = list(), # No dependencies - key test case
+        unserialize_function = "pickle.load",
+        type = "rxp_py"
+      )
+    )
+  )
+  
+  # Write DAG to temporary file
+  dag_file <- tempfile(fileext = ".json")
+  jsonlite::write_json(dag_content, dag_file, pretty = TRUE)
+  
+  # Create flat pipeline with placeholders
+  flat_pipeline <- c(
+    "  no_deps_r = makeRDerivation {",
+    "    name = \"no_deps_r\";",
+    "    buildInputs = defaultBuildInputs;",
+    "    configurePhase = defaultConfigurePhase;",
+    "    buildPhase = ''",
+    "      Rscript -e \"",
+    "        source('libraries.R')",
+    "        # RIXPRESS_LOAD_DEPENDENCIES_HERE",
+    "        no_deps_r <- list(x = 1)",
+    "        saveRDS(no_deps_r, 'no_deps_r')\"",
+    "    '';",
+    "  };",
+    "",
+    "  with_deps_r = makeRDerivation {",
+    "    name = \"with_deps_r\";", 
+    "    buildInputs = defaultBuildInputs;",
+    "    configurePhase = defaultConfigurePhase;",
+    "    buildPhase = ''",
+    "      Rscript -e \"",
+    "        source('libraries.R')",
+    "        # RIXPRESS_LOAD_DEPENDENCIES_HERE", 
+    "        with_deps_r <- no_deps_r$x + 1",
+    "        saveRDS(with_deps_r, 'with_deps_r')\"",
+    "    '';",
+    "  };",
+    "",
+    "  no_deps_py = makePyDerivation {",
+    "    name = \"no_deps_py\";",
+    "    buildInputs = defaultBuildInputs;",
+    "    configurePhase = defaultConfigurePhase;",
+    "    buildPhase = ''",
+    "      python -c \"",
+    "exec(open('libraries.py').read())",
+    "# RIXPRESS_PY_LOAD_DEPENDENCIES_HERE",
+    "exec('no_deps_py = [1, 2, 3]')",
+    "with open('no_deps_py', 'wb') as f: pickle.dump(globals()['no_deps_py'], f)",
+    "\"",
+    "    '';",
+    "  };"
+  )
+  
+  # Call gen_pipeline function
+  result_pipeline <- gen_pipeline(dag_file, flat_pipeline)
+  result_str <- paste(result_pipeline, collapse = "\n")
+  
+  # Test 1: Derivations with no dependencies should have placeholders replaced with empty strings
+  # Check that placeholders are removed
+  expect_false(grepl("# RIXPRESS_LOAD_DEPENDENCIES_HERE", result_str))
+  expect_false(grepl("# RIXPRESS_PY_LOAD_DEPENDENCIES_HERE", result_str))
+  
[…]
