### 2025-08-27 08f878b
much better handling of derivs deps
