### 2025-08-27 7738400
better handling of serialization functions
