+++ R/rxp_io.R
@@ -5,26 +5,10 @@
 #' @param user_functions Character vector of user-defined function script paths.
 #' @return Character vector without empty strings.
 clean_user_functions <- function(user_functions) {
-  if (is.null(user_functions)) return(character(0))
+  if (is.null(user_functions)) {
+    return(character(0))
+  }
   user_functions[nzchar(user_functions)]
-}
-
-#' Build copy commands for user functions
-#'
-#' Generates shell commands to copy user-defined function scripts into the
-#' build sandbox.
-#'
-#' @param user_functions Character vector of script file names.
-#' @return A string of `cp` commands, suitable for embedding in a build phase.
-build_user_copy_cmd <- function(user_functions) {
-  if (length(user_functions) == 0) return("")
-  paste0(
-    paste(
-      sprintf("cp ${./%s} %s", user_functions, user_functions),
-      collapse = "\n      "
-    ),
-    "\n      "
-  )
 }
 
 #' Build code import/source commands for user functions
@@ -35,14 +19,19 @@ build_user_copy_cmd <- function(user_functions) {
 #' @param lang Language string, `"R"` or `"Python"`.
 #' @return A string of import/source statements.
 build_user_code_cmd <- function(user_functions, lang) {
-  if (length(user_functions) == 0) return("")
+  if (length(user_functions) == 0) {
+    return("")
+  }
+  # Always prepend input_folder/
+  files <- paste0("input_folder/", user_functions)
   fmt <- switch(
     lang,
-    R      = "source('%s')",
+    R = "source('%s')",
     Py = "exec(open('%s').read())",
+    jl = "include('%s')",
     stop("Unsupported lang: ", lang)
   )
-  paste(sprintf(fmt, user_functions), collapse = "\n")
+  paste(sprintf(fmt, files), collapse = "\n")
 }
 
 #' Build environment variable export commands
@@ -53,11 +42,17 @@ build_user_code_cmd <- function(user_functions, lang) {
 #' @param env_var Named list of environment variables.
 #' @return A string of `export` commands with line breaks, or `""` if none.
 build_env_exports <- function(env_var) {
-  if (is.null(env_var) || length(env_var) == 0) return("")
-  lines <- vapply(names(env_var), function(var) {
-    sprintf("export %s=%s", var, env_var[[var]])
-  }, character(1))
-  paste0(paste(lines, collapse = "\n      "), "\n      ")
+  if (is.null(env_var) || length(env_var) == 0) {
+    return("")
+  }
+  lines <- vapply(
+    names(env_var),
+    function(var) {
+      sprintf("export %s=%s", var, env_var[[var]])
+    },
+    character(1)
+  )
+  paste0(paste(lines, collapse = "\n"), "\n")
 }
 
 #' Build Nix src part
@@ -75,23 +70,22 @@ build_src_part <- function(actual_path, user_functions) {
       system(paste("nix-prefetch-url", shQuote(actual_path)), intern = TRUE),
       error = function(e) stop("Failed to run nix-prefetch-url: ", e$message)
     )
-    if (length(hash) == 0 || hash == "")
+    if (length(hash) == 0 || hash == "") {
       stop("nix-prefetch-url did not return a hash for URL: ", actual_path)
-
+    }
     sprintf(
       "defaultPkgs.fetchurl {\n      url = \"%s\";\n      sha256 = \"%s\";\n    }",
       actual_path,
       trimws(hash[1])
     )
-  } else if (length(user_functions) > 0) {
-    fileset_parts <- c(actual_path, user_functions)
-    fileset_nix <- paste0("./", fileset_parts, collapse = " ")
+  } else {
[…]
