+++ R/rxp_ga.R
@@ -46,6 +46,7 @@ rxp_ga <- function() {
     message("GitHub Actions workflow file saved to: ", path)
   }
 
-  if (identical(Sys.getenv("TESTTHAT"), "true"))
+  if (identical(Sys.getenv("TESTTHAT"), "true")) {
     paste0(path, "/run-rxp-pipeline.yaml")
+  }
 }

[…]
