+++ R/rxp_io.R
@@ -1,3 +1,22 @@
+# Constants for language configuration
+LANG_CONFIG <- list(
+  R = list(
+    source_cmd = "source('%s')",
+    derivation_type = "R",
+    output_ext = "RDS"
+  ),
+  Py = list(
+    source_cmd = "exec(open('%s').read())",
+    derivation_type = "Py",
+    output_ext = "pickle"
+  ),
+  Jl = list(
+    source_cmd = "include('%s')",
+    derivation_type = "Jl",
+    output_ext = "jld2"
+  )
+)
+
 #' Clean user functions vector
 #'
 #' Removes empty character entries from the provided `user_functions` vector.
@@ -16,22 +35,20 @@ clean_user_functions <- function(user_functions) {
 #' Creates language-specific commands to load user-defined function scripts.
 #'
 #' @param user_functions Character vector of script file names.
-#' @param lang Language string, `"R"`, `"Python"` and `"Julia"`.
+#' @param lang Language string, `"R"`, `"Py"`, or `"Jl"`.
 #' @return A string of import/source statements.
 build_user_code_cmd <- function(user_functions, lang) {
   if (length(user_functions) == 0) {
     return("")
   }
-  # Always prepend input_folder/
-  files <- paste0("input_folder/", user_functions)
-  fmt <- switch(
-    lang,
-    R = "source('%s')",
-    Py = "exec(open('%s').read())",
-    Jl = "include('%s')",
+
+  if (!lang %in% names(LANG_CONFIG)) {
     stop("Unsupported lang: ", lang)
-  )
-  paste(sprintf(fmt, files), collapse = "\n")
+  }
+
+  files <- file.path("input_folder", user_functions)
+  commands <- sprintf(LANG_CONFIG[[lang]]$source_cmd, files)
+  paste(commands, collapse = "\n")
 }
 
 #' Build environment variable export commands
@@ -45,14 +62,9 @@ build_env_exports <- function(env_var) {
   if (is.null(env_var) || length(env_var) == 0) {
     return("")
   }
-  lines <- vapply(
-    names(env_var),
-    function(var) {
-      sprintf("export %s=%s", var, env_var[[var]])
-    },
-    character(1)
-  )
-  paste0(paste(lines, collapse = "\n"), "\n")
+
+  exports <- sprintf("export %s=%s", names(env_var), env_var)
+  paste0(paste(exports, collapse = "\n"), "\n")
 }
 
 #' Build Nix src part
@@ -61,32 +73,57 @@ build_env_exports <- function(env_var) {
 #' (using `nix-prefetch-url`) and local paths (with optional user function
 #' scripts).
 #'
-#' @param actual_path Character path to file or directory.
+#' @param path Character path to file or directory.
 #' @param user_functions Character vector of additional files to include.
 #' @return A string with the `src` Nix expression.
-build_src_part <- function(actual_path, user_functions) {
-  if (grepl("^https?://", actual_path)) {
-    hash <- tryCatch(
-      system(paste("nix-prefetch-url", shQuote(actual_path)), intern = TRUE),
-      error = function(e) stop("Failed to run nix-prefetch-url: ", e$message)
-    )
-    if (length(hash) == 0 || hash == "") {
-      stop("nix-prefetch-url did not return a hash for URL: ", actual_path)
-    }
-    sprintf(
-      "defaultPkgs.fetchurl {\n      url = \"%s\";\n      sha256 = \"%s\";\n    }",
-      actual_path,
-      trimws(hash[1])
-    )
+build_src_part <- function(path, user_functions = character(0)) {
+  if (is_remote_url(path)) {
+    return(build_remote_src(path))
   } else {
[…]
