### 2025-08-28 cba8898
documentation and testing
