+++ man/rxp_file.Rd
@@ -1,12 +1,22 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rxp_io.R
-\name{rxp_r_file}
-\alias{rxp_r_file}
-\title{Creates a Nix expression that reads in a file (or folder of data) using R.}
+\name{rxp_file}
+\alias{rxp_file}
+\title{Generic Nix expression builder for R, Python, and Julia data readers}
 \usage{
-rxp_r_file(...)
+rxp_file(
+  lang,
+  name,
+  path,
+  read_function,
+  user_functions = "",
+  nix_env = "default.nix",
+  env_var = NULL
+)
 }
 \arguments{
+\item{lang}{\code{"R"}, \code{"Py"} or \code{"Jl"}.}
+
 \item{name}{Symbol, the name of the derivation.}
 
 \item{path}{Character, file or folder path to include.}
@@ -23,16 +33,7 @@ rxp_r_file(...)
 An object of class \code{rxp_derivation}.
 }
 \description{
-Creates a Nix expression that reads in a file (or folder of data) using R.
+Creates a Nix derivation that reads a file or folder of data using R,
+Python, or Julia. Handles user-defined functions, environment variables, and Nix
+environment specification.
 }
-\seealso{
-Other derivations: 
-\code{\link{rxp_jl}()},
-\code{\link{rxp_jl_file}()},
-\code{\link{rxp_py}()},
-\code{\link{rxp_py_file}()},
-\code{\link{rxp_qmd}()},
-\code{\link{rxp_r}()},
-\code{\link{rxp_rmd}()}
-}
-\concept{derivations}
