+++ man/build_remote_src.Rd
@@ -1,0 +1,17 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rxp_io.R
+\name{build_remote_src}
+\alias{build_remote_src}
+\title{Build remote source configuration}
+\usage{
+build_remote_src(url)
+}
+\arguments{
+\item{url}{Character URL}
+}
+\value{
+Character Nix expression
+}
+\description{
+Build remote source configuration
+}
