+++ man/rxp_r2py.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/derivs.R
+% Please edit documentation in R/rxp_io.R
 \name{rxp_r2py}
 \alias{rxp_r2py}
 \title{Transfer R object into a Python session.}
@@ -15,7 +15,7 @@ rxp_r2py(name, expr, nix_env = "default.nix")
 "default.nix".}
 }
 \value{
-An object of class derivation which inherits from lists.
+An object of class \code{rxp_derivation}.
 }
 \description{
 Transfer R object into a Python session.
@@ -23,11 +23,6 @@ Transfer R object into a Python session.
 \details{
 \code{rxp_r2py(my_obj, my_r_object)} saves an R object to a Python pickle
 using \code{reticulate::py_save_object()}.
-}
-\examples{
-\dontrun{
-  rxp_r2py(my_obj, my_r_object)
-}
 }
 \seealso{
 Other interop functions: 
