+++ man/build_user_code_cmd.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rxp_io.R
+\name{build_user_code_cmd}
+\alias{build_user_code_cmd}
+\title{Build code import/source commands for user functions}
+\usage{
+build_user_code_cmd(user_functions, lang)
+}
+\arguments{
+\item{user_functions}{Character vector of script file names.}
+
+\item{lang}{Language string, \code{"R"}, \code{"Py"}, or \code{"Jl"}.}
+}
+\value{
+A string of import/source statements.
+}
+\description{
+Creates language-specific commands to load user-defined function scripts.
+}
