+++ R/rxp_io.R
@@ -78,10 +78,50 @@ build_env_exports <- function(env_var) {
 #' @return A string with the `src` Nix expression.
 build_src_part <- function(path, user_functions = character(0)) {
   if (is_remote_url(path)) {
-    return(build_remote_src(path))
+    if (length(user_functions) > 0) {
+      return(build_hybrid_src(path, user_functions))
+    } else {
+      return(build_remote_src(path))
+    }
   } else {
     return(build_local_src(path, user_functions))
   }
+}
+
+#' Build hybrid source configuration for remote URL + local user functions
+#' @param url Character URL
+#' @param user_functions Character vector of additional files
+#' @return Character Nix expression
+build_hybrid_src <- function(url, user_functions) {
+  hash <- tryCatch(
+    system(paste("nix-prefetch-url", shQuote(url)), intern = TRUE),
+    error = function(e) stop("Failed to run nix-prefetch-url: ", e$message)
+  )
+
+  if (length(hash) == 0 || hash == "") {
+    stop("nix-prefetch-url did not return a hash for URL: ", url)
+  }
+
+  remote_filename <- basename(url)
+  
+  # Create the local fileset for user functions
+  user_files_list <- paste0("./", user_functions, collapse = " ")
+  local_src_var <- sprintf(
+    "localSrc = defaultPkgs.lib.fileset.toSource {\n        root = ./.;\n        fileset = defaultPkgs.lib.fileset.unions [ %s ];\n      }",
+    user_files_list
+  )
+  
+  # Create copy commands for user functions
+  user_files_cp <- paste(sprintf("cp ${localSrc}/%s $out/", user_functions), collapse = "\n      ")
+  
+  sprintf(
+    "(let\n      %s;\n    in defaultPkgs.runCommand \"combined-src\" {} ''\n      mkdir -p $out\n      cp ${defaultPkgs.fetchurl {\n        url = \"%s\";\n        sha256 = \"%s\";\n      }} $out/%s\n      %s\n    '')",
+    local_src_var,
+    url,
+    trimws(hash[1]),
+    remote_filename,
+    user_files_cp
+  )
 }
 
 #' Check if path is a remote URL
@@ -195,15 +235,27 @@ build_language_commands <- function(
 #' @param user_code Source/import statements for user functions.
 #' @param out_name Name of the output object (RDS/pickle file).
 #' @param path Input path (file or folder).
+#' @param user_functions Character vector of user function files.
 #' @return A string with the build phase commands.
-build_phase <- function(lang, read_func, user_code, out_name, path) {
-  # For remote URLs we want the local filename that $src will copy into input_folder.
-  # e.g. fetchurl of "https://.../mtcars.csv" will be copied to input_folder/mtcars.csv
+build_phase <- function(lang, read_func, user_code, out_name, path, user_functions = character(0)) {
   if (is_remote_url(path)) {
     rel_name <- basename(path)
     rel_path <- rel_name
-    copy_line <- paste0("cp -r $src ", rel_path)
+    
+    if (length(user_functions) > 0) {
+      # Hybrid case: everything is in $src, just organize it
+      copy_line <- paste(
+        "cp $src/* .",
+        "mkdir -p input_folder",
+        paste(sprintf("cp %s input_folder/", basename(user_functions)), collapse = "\n"),
+        sep = "\n"
+      )
+    } else {
+      # Pure remote case
+      copy_line <- paste0("cp -r $src ", rel_path)
+    }
   } else {
+    # Local case
     rel_name <- path
     rel_path <- file.path("input_folder", rel_name)
     copy_line <- "cp -r $src input_folder"
@@ -297,7 +349,7 @@ rxp_file <- function(
   # Build components
[…]
