+++ man/process_read_function.Rd
@@ -19,3 +19,4 @@ Character string
 \description{
 Process read function for different languages
 }
+\keyword{internal}
