+++ man/build_language_commands.Rd
@@ -23,3 +23,4 @@ Character vector of commands
 \description{
 Build language-specific execution commands
 }
+\keyword{internal}
