### 2025-08-28 c72c8db
moar internal keywords
