+++ R/derivs.R
@@ -5,6 +5,7 @@
 #' @param base Character, base name for buildInputs and configurePhase
 #' @param build_phase Character, the build phase commands
 #' @param derivation_type Character, one of "R", "Py", "Jl", "Qmd", "Rmd"
+#' @param build Logical, whether to actually build the derivation (default TRUE)
 #' @return Character string with the formatted Nix derivation
 #' @keywords internal
 make_derivation_snippet <- function(
@@ -12,8 +13,20 @@ make_derivation_snippet <- function(
   src_snippet,
   base,
   build_phase,
-  derivation_type
+  derivation_type,
+  build = TRUE
 ) {
+  # If build is FALSE, create a dummy derivation that just creates an empty output
+  if (!build) {
+    return(sprintf(
+      "  %s = defaultPkgs.runCommand \"%s\" {} \"\n    mkdir -p $out\n    echo 'Build skipped for %s' > $out/SKIPPED\n  \";",
+      out_name,
+      out_name,
+      out_name
+    ))
+  }
+
+  # Normal build logic for when build = TRUE
   # Determine the derivation function based on type
   derivation_func <- switch(
     derivation_type,
@@ -78,6 +91,9 @@ make_derivation_snippet <- function(
 #'   environment variables to set before running the R script, e.g.,
 #'   `c("CMDSTAN" = "${defaultPkgs.cmdstan}/opt/cmdstan)"`.
 #'   Each entry will be added as an export statement in the build phase.
+#' @param build Logical, defaults to TRUE. If TRUE, the derivation will be built
+#'   when the pipeline runs. If FALSE, the derivation definition is created but
+#'   the build is skipped.
 #' @details At a basic level, `rxp_r(mtcars_am, filter(mtcars, am == 1))` is
 #'   equivalent to `mtcars_am <- filter(mtcars, am == 1)`. `rxp_r()` generates the
 #'   required Nix boilerplate to output a so-called "derivation" in Nix jargon.
@@ -88,6 +104,13 @@ make_derivation_snippet <- function(
 #' @examples \dontrun{
 #'   # Basic usage
 #'   rxp_r(name = filtered_mtcars, expr = filter(mtcars, am == 1))
+#'
+#'   # Skip building this derivation
+#'   rxp_r(
+#'     name = turtles,
+#'     expr = occurrence(species, geometry = atlantic),
+#'     build = FALSE
+#'   )
 #'
 #'   # Serialize object using qs
 #'   rxp_r(
@@ -111,7 +134,8 @@ rxp_r <- function(
   nix_env = "default.nix",
   serialize_function = NULL,
   unserialize_function = NULL,
-  env_var = NULL
+  env_var = NULL,
+  build = TRUE
 ) {
   out_name <- deparse1(substitute(name))
   expr_str <- deparse1(substitute(expr))
@@ -260,7 +284,8 @@ rxp_r <- function(
     nix_env = nix_env,
     serialize_function = serialize_str,
     unserialize_function = unserialize_str,
-    env_var = env_var
+    env_var = env_var,
+    build = build
   ) |>
     structure(class = "rxp_derivation")
 }
@@ -292,6 +317,9 @@ rxp_r <- function(
 #'   environment variables
 #'   before running the Python script, e.g., c(PYTHONPATH = "/path/to/modules").
 #'   Each entry will be added as an export statement in the build phase.
+#' @param build Logical, defaults to TRUE. If TRUE, the derivation will be built
+#'   when the pipeline runs. If FALSE, the derivation definition is created but
+#'   the build is skipped.
[…]
