+++ man/rxp_rmd.Rd
@@ -11,7 +11,7 @@ rxp_rmd(
   nix_env = "default.nix",
   params = NULL,
   env_var = NULL,
-  build = TRUE
+  noop_build = FALSE
 )
 }
 \arguments{
@@ -34,9 +34,10 @@ to set before running the R Markdown render command, e.g., c(RSTUDIO_PANDOC
 = "/path/to/pandoc"). Each entry will be added as an export statement in
 the build phase.}
 
-\item{build}{Logical, defaults to TRUE. If TRUE, the derivation will be built
-when the pipeline runs. If FALSE, the derivation definition is created but
-the build is skipped.}
+\item{noop_build}{Logical, defaults to FALSE. If TRUE, the derivation
+produces a no-op build (a stub output with no actual build steps). Any
+downstream derivations depending on a no-op build will themselves also
+become no-op builds.}
 }
 \value{
 An object of class derivation which inherits from lists.
@@ -62,7 +63,7 @@ To include objects built in the pipeline,
   rxp_rmd(
     name = draft_report,
     rmd_file = "draft.Rmd",
-    build = FALSE
+    noop_build = TRUE
   )
 }
 }

[…]
