+++ R/rxp_make.R
@@ -5,15 +5,6 @@
 #' @keywords internal
 #' @noRd
 .rxp_validate_verbose <- function(verbose) {
-  if (is.logical(verbose)) {
-    warning(
-      "logical values for 'verbose' are deprecated. ",
-      "Use integer values instead: verbose = 0 (FALSE) or verbose = 1 (TRUE).",
-      call. = FALSE
-    )
-    return(as.integer(verbose))
-  }
-
   if (!is.numeric(verbose) || length(verbose) != 1) {
     stop("verbose must be a single numeric or integer value", call. = FALSE)
   }
@@ -24,7 +15,110 @@
     stop("rxp_make(): verbose must be a non-negative integer", call. = FALSE)
   }
 
+  if (verbose > 5L) {
+    warning(
+      sprintf(
+        "Argument 'verbose' (%d) exceeds the maximum of 5; using 5.",
+        verbose
+      ),
+      call. = FALSE
+    )
+    verbose <- 5L
+  }
+
   verbose
+}
+
+#' Parse Nix build output to track derivation progress
+#'
+#' @param line Output line from nix-store
+#' @param derivation_names Vector of derivation names to track
+#' @return List with status info, or NULL if not a relevant line
+#' @keywords internal
+#' @noRd
+#' Parse Nix build output to track derivation progress
+#'
+#' @param line Output line from nix-store
+#' @param derivation_names Vector of derivation names to track
+#' @return List with status info, or NULL if not a relevant line
+#' @keywords internal
+#' @noRd
+.rxp_parse_build_line <- function(line, derivation_names) {
+  # Handle final store paths (when derivations are already built)
+  if (grepl("^/nix/store/.*-[^/]+$", line)) {
+    for (deriv_name in derivation_names) {
+      if (grepl(paste0("-", deriv_name, "$"), line)) {
+        # Skip internal derivations from progress display
+        if (deriv_name == "all-derivations") {
+          return(NULL)
+        }
+        return(list(
+          derivation = deriv_name,
+          status = "completed",
+          message = line,
+          from_store = TRUE
+        ))
+      }
+    }
+  }
+
+  # Handle dispatch: "building '/nix/store/...-derivation_name.drv'..."
+  if (grepl("building '/nix/store/.*\\.drv'\\.\\.\\.", line)) {
+    for (deriv_name in derivation_names) {
+      if (grepl(paste0("-", deriv_name, "\\.drv"), line)) {
+        # Skip internal derivations from progress display
+        if (deriv_name == "all-derivations") {
+          return(NULL)
+        }
+        return(list(
+          derivation = deriv_name,
+          status = "dispatched",
+          message = line,
+          from_store = FALSE
+        ))
+      }
+    }
+  }
+
+  # Handle direct build failures: "error: builder for '/nix/store/...-derivation_name.drv' failed"
+  if (grepl("error: builder for.*failed", line)) {
+    for (deriv_name in derivation_names) {
+      if (grepl(paste0("-", deriv_name, "\\.drv"), line)) {
+        return(list(
+          derivation = deriv_name,
+          status = "errored",
+          message = line,
+          from_store = FALSE
+        ))
+      }
+    }
+  }
+
+  # Handle dependency failures: "error: X dependencies of derivation '/nix/store/...-derivation_name.drv' failed"
+  if (grepl("error: .* dependencies of derivation.*failed", line)) {
+    for (deriv_name in derivation_names) {
+      if (grepl(paste0("-", deriv_name, "\\.drv"), line)) {
+        return(list(
+          derivation = deriv_name,
+          status = "errored",
+          message = line,
[…]
