+++ data-raw/gen_pipeline.R
@@ -17,7 +17,8 @@ d2 <- rxp_r(
   mtcars_head,
   my_head(mtcars_am, 100),
   user_functions = "my_head.R",
-  nix_env = "default.nix"
+  nix_env = "default.nix",
+  serialize_function = write.csv
 )
 
 d3 <- rxp_r(
@@ -25,10 +26,16 @@ d3 <- rxp_r(
   my_tail(mtcars_head),
   user_functions = "my_tail.R",
   nix_env = "default.nix",
-  noop_build = TRUE
+  serialize_function = qs::qsave,
+  unserialize_function = read.csv
 )
 
-d4 <- rxp_r(mtcars_mpg, select(mtcars_tail, mpg), nix_env = "default2.nix")
+d4 <- rxp_r(
+  mtcars_mpg,
+  full_join(mtcars_tail, mtcars_head),
+  nix_env = "default2.nix",
+  unserialize_function = c(mtcars_tail = "qs::qread", mtcars_head = "read.csv")
+)
 
 doc <- rxp_qmd(
   page,
