+++ R/derivs.R
@@ -84,12 +84,13 @@ make_derivation_snippet <- function(
 #'   instance, for `{keras}` models, use `keras::save_model_hdf5()` to capture
 #'   the full model (architecture, weights, training config, optimiser state,
 #'   etc.).
-#' @param unserialize_function Function, defaults to NULL. A function used to
-#'   unserialize objects transferred between derivations. By default,
-#'   `readRDS()` is used, but this may produce unexpected results with complex
-#'   objects like machine learning models. For example, if the parent derivation
-#'   used `keras::save_model_hdf5()` to serialize a model, this derivation
-#'   should use `keras::load_model_hdf5()` to load it correctly.
+#' @param unserialize_function Function, character, or named vector/list,
+#'   defaults to NULL. Can be:
+#'   - A single function/string to unserialize all upstream objects (e.g., `readRDS`)
+#'   - A named vector/list where names are upstream dependency names and values
+#'     are their specific unserialize functions (e.g.,
+#'     `c(mtcars_tail = "qs::qread", mtcars_head = "read.csv")`)
+#'   By default, `readRDS()` is used.
 #' @param env_var Character vector, defaults to NULL. A named vector of
 #'   environment variables to set before running the R script, e.g.,
 #'   `c("CMDSTAN" = "${defaultPkgs.cmdstan}/opt/cmdstan)"`.
@@ -158,13 +159,70 @@ rxp_r <- function(
     serialize_str <- deparse1(serialize_expr)
   }
 
+  # Handle unserialize_function - can be single value or named vector/list
   unserialize_expr <- substitute(unserialize_function)
   if (identical(unserialize_expr, quote(NULL))) {
     unserialize_str <- "readRDS"
-  } else if (is.character(unserialize_expr)) {
-    unserialize_str <- unserialize_expr
   } else {
-    unserialize_str <- deparse1(unserialize_expr)
+    # Check if it's a call to c() with named arguments
+    if (
+      is.call(unserialize_expr) && identical(unserialize_expr[[1]], quote(c))
+    ) {
+      # It's a c() call - check if it has names
+      call_names <- names(unserialize_expr)
+      if (!is.null(call_names) && any(nzchar(call_names[-1]))) {
+        # Has named arguments (skip first element which is the function name 'c')
+        # Extract as a named list to preserve names in JSON
+        unserialize_str <- list()
+        for (i in 2:length(unserialize_expr)) {
+          if (nzchar(call_names[i])) {
+            # Get the value as a string
+            val <- unserialize_expr[[i]]
+            if (is.character(val)) {
+              unserialize_str[[call_names[i]]] <- val
+            } else {
+              unserialize_str[[call_names[i]]] <- as.character(val)
+            }
+          }
+        }
+        # Convert to a named list explicitly
+        unserialize_str <- as.list(unserialize_str)
+      } else {
+        # No names, treat as single value
+        unserialize_str <- deparse1(unserialize_expr)
+      }
+    } else if (is.character(unserialize_expr)) {
+      # Direct character value
+      unserialize_str <- unserialize_expr
+    } else {
+      # Try to evaluate it to see if it's a pre-existing named vector
+      tryCatch(
+        {
+          unserialize_val <- eval(unserialize_expr, envir = parent.frame())
+          if (
+            !is.null(names(unserialize_val)) &&
+              length(names(unserialize_val)) > 0
+          ) {
+            # Convert named vector to named list for JSON preservation
+            unserialize_str <- as.list(setNames(
+              unserialize_val,
+              names(unserialize_val)
+            ))
+          } else {
+            # Single value
+            if (is.character(unserialize_val)) {
+              unserialize_str <- unserialize_val
+            } else {
+              unserialize_str <- deparse1(unserialize_expr)
+            }
+          }
+        },
[…]
