+++ R/derivs.R
@@ -61,6 +61,273 @@ make_derivation_snippet <- function(
   )
 }
 
+
+#' Helper function to parse unserialize_function parameter
+#'
+#' @param unserialize_expr Expression from substitute(unserialize_function)
+#' @param default_func Default function to use if NULL
+#' @param parent_env Parent environment for evaluation
+#' @return Processed unserialize string or list
+#' @keywords internal
+parse_unserialize_function <- function(
+  unserialize_expr,
+  default_func,
+  parent_env = parent.frame()
+) {
+  if (identical(unserialize_expr, quote(NULL))) {
+    return(default_func)
+  }
+
+  # Check if it's a call to c() with named arguments
+  if (is.call(unserialize_expr) && identical(unserialize_expr[[1]], quote(c))) {
+    call_names <- names(unserialize_expr)
+    if (!is.null(call_names) && any(nzchar(call_names[-1]))) {
+      # Has named arguments - extract as named list
+      unserialize_str <- list()
+      for (i in 2:length(unserialize_expr)) {
+        if (nzchar(call_names[i])) {
+          val <- unserialize_expr[[i]]
+          unserialize_str[[call_names[i]]] <- if (is.character(val)) {
+            val
+          } else {
+            as.character(val)
+          }
+        }
+      }
+      return(as.list(unserialize_str))
+    } else {
+      # No names, treat as single value
+      if (length(unserialize_expr) > 1) {
+        val <- unserialize_expr[[2]]
+        return(if (is.character(val)) val else as.character(val))
+      }
+      return(deparse1(unserialize_expr))
+    }
+  }
+
+  if (is.character(unserialize_expr)) {
+    return(unserialize_expr)
+  }
+
+  # Try to evaluate it
+  tryCatch(
+    {
+      unserialize_val <- eval(unserialize_expr, envir = parent_env)
+      if (
+        !is.null(names(unserialize_val)) && length(names(unserialize_val)) > 0
+      ) {
+        as.list(setNames(unserialize_val, names(unserialize_val)))
+      } else {
+        if (is.character(unserialize_val)) {
+          unserialize_val
+        } else {
+          deparse1(unserialize_expr)
+        }
+      }
+    },
+    error = function(e) {
+      deparse1(unserialize_expr)
+    }
+  )
+}
+
+#' Build environment variable export statements
+#'
+#' @param env_var Named vector of environment variables
+#' @param indent Number of spaces for indentation
+#' @return String with export statements
+#' @keywords internal
+build_env_exports <- function(env_var, indent = 6) {
+  if (is.null(env_var) || length(env_var) == 0) {
+    return("")
+  }
+
+  indent_str <- paste(rep(" ", indent), collapse = "")
+  exports <- vapply(
+    names(env_var),
+    function(var) sprintf("export %s=%s", var, env_var[[var]]),
+    character(1)
+  )
+
+  paste0(
+    paste(paste0(indent_str, exports), collapse = "\n"),
+    "\n",
+    indent_str
+  )
+}
+
+#' Build copy commands for files
+#'
+#' @param files Character vector of files to copy
+#' @param recursive Logical, use cp -r if TRUE
+#' @param indent Number of spaces for indentation
+#' @return String with copy commands
+#' @keywords internal
+build_copy_commands <- function(files, recursive = TRUE, indent = 6) {
+  if (is.null(files) || !any(nzchar(files))) {
+    return("")
+  }
+
+  files_clean <- files[nzchar(files)]
+  if (length(files_clean) == 0) {
+    return("")
+  }
+
+  indent_str <- paste(rep(" ", indent), collapse = "")
[…]
