+++ man/parse_unserialize_function.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/derivs.R
+\name{parse_unserialize_function}
+\alias{parse_unserialize_function}
+\title{Helper function to parse unserialize_function parameter}
+\usage{
+parse_unserialize_function(
+  unserialize_expr,
+  default_func,
+  parent_env = parent.frame()
+)
+}
+\arguments{
+\item{unserialize_expr}{Expression from substitute(unserialize_function)}
+
+\item{default_func}{Default function to use if NULL}
+
+\item{parent_env}{Parent environment for evaluation}
+}
+\value{
+Processed unserialize string or list
+}
+\description{
+Helper function to parse unserialize_function parameter
+}
+\keyword{internal}
