+++ man/build_env_exports.Rd
@@ -1,15 +1,21 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rxp_io.R
+% Please edit documentation in R/derivs.R, R/rxp_io.R
 \name{build_env_exports}
 \alias{build_env_exports}
-\title{Build environment variable export commands}
+\title{Build environment variable export statements}
 \usage{
+build_env_exports(env_var)
+
 build_env_exports(env_var)
 }
 \arguments{
 \item{env_var}{Named list of environment variables.}
+
+\item{indent}{Number of spaces for indentation}
 }
 \value{
+String with export statements
+
 A string of \code{export} commands with line breaks, or \code{""} if none.
 }
 \description{
