+++ man/extract_rxp_matches.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/derivs.R
+\name{extract_rxp_matches}
+\alias{extract_rxp_matches}
+\title{Helper function to extract rxp_read/rxp_load matches from content}
+\usage{
+extract_rxp_matches(content_str, func_name)
+}
+\arguments{
+\item{content_str}{String content of the document}
+
+\item{func_name}{Function name to search for ("rxp_read" or "rxp_load")}
+}
+\value{
+Data frame with match information
+}
+\description{
+Helper function to extract rxp_read/rxp_load matches from content
+}
+\keyword{internal}
