+++ man/create_derivation.Rd
@@ -1,0 +1,50 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/derivs.R
+\name{create_derivation}
+\alias{create_derivation}
+\title{Create derivation object}
+\usage{
+create_derivation(
+  name,
+  snippet,
+  type,
+  additional_files = "",
+  user_functions = "",
+  nix_env = "default.nix",
+  serialize_function = NULL,
+  unserialize_function = NULL,
+  env_var = NULL,
+  noop_build = FALSE,
+  ...
+)
+}
+\arguments{
+\item{name}{Derivation name}
+
+\item{snippet}{Nix snippet}
+
+\item{type}{Derivation type}
+
+\item{additional_files}{Additional files}
+
+\item{user_functions}{User function files}
+
+\item{nix_env}{Nix environment}
+
+\item{serialize_function}{Serialize function}
+
+\item{unserialize_function}{Unserialize function}
+
+\item{env_var}{Environment variables}
+
+\item{noop_build}{No-op build flag}
+
+\item{...}{Additional fields for specific derivation types}
+}
+\value{
+Derivation object with class "rxp_derivation"
+}
+\description{
+Create derivation object
+}
+\keyword{internal}
