+++ man/build_source_commands.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/derivs.R
+\name{build_source_commands}
+\alias{build_source_commands}
+\title{Build source/import commands for user functions}
+\usage{
+build_source_commands(user_functions, lang, indent = 8)
+}
+\arguments{
+\item{user_functions}{Character vector of function files}
+
+\item{lang}{Language: "R", "Py", or "Jl"}
+
+\item{indent}{Number of spaces for indentation}
+}
+\value{
+String with source/import commands
+}
+\description{
+Build source/import commands for user functions
+}
+\keyword{internal}
