### 2025-08-30 dbed444
Improve clarity and structure of README.md
Refactor README for clarity and conciseness, improving language and structure throughout.
