+++ README.md
@@ -16,18 +16,17 @@ below:
 <img src="https://raw.githubusercontent.com/b-rodrigues/rixpress/refs/heads/main/video_thumbnail.png" alt="Video Thumbnail" style="width:100%; max-width:560px; height:auto; display:block; margin:0 auto;">
 </a>
 
-This R package provides a framework for building multilanguage
-reproducible analytical pipelines by leveraging `Nix`’s build automation
-capabilities. One of the design goals of `{rixpress}` is to mimic the
-user experience of the `{targets}` package, and is thus heavily inspired
-by it. It builds upon the `{rix}` package, which provides helper
-function to define reproducible development environments as code using
-`Nix`, ensuring the pipeline runs in a fully reproducible `Nix`-managed
-environment. `{rixpress}` only requires users to write the pipeline
-using familiar R code.
+`{rixpress}` provides a framework for building multilanguage reproducible
+analytical pipelines by leveraging Nix’s build automation capabilities.
+One of the design goals of `{rixpress}` is to mimic the user experience of
+the `{targets}` package, and it is heavily inspired by that workflow. It
+builds on the `{rix}` package, which provides helper functions to define
+reproducible development environments as code using Nix, ensuring the
+pipeline runs in a fully reproducible Nix-managed environment. `{rixpress}`
+only requires users to write the pipeline using familiar R code.
 
 `rixpress` focuses on “micropipelines”: pipelines executed on a single
-machine for small-to-medium projects.
+machine for small-to-medium sized projects.
 
 For example, this R script defines a list of *derivations* defined by
 functions prefixed with `rxp_*()`, which is then passed to `rxp_populate()`:
@@ -70,39 +69,37 @@ list(
   rxp_populate()
 ```
 
-Running `rxp_populate()` generates a `pipeline.nix` file, which contains all
-the build instructions of all the derivations and final outputs as `Nix`
-code. It is possible to define derivations that run Python or Julia code
-as well, and objects can be passed to and from R or Python by using
-`rxp_py2r()` and `ryp_r2py()`, or by using a common serialization format
-such as JSON. By default, calling `rxp_populate()` also builds the pipeline,
-but it’s possible to only generate the `pipeline.nix` file and then
-build the pipeline at a later stage using:
+Running `rxp_populate()` generates a `pipeline.nix` file, which contains
+the build instructions for all derivations and final outputs expressed
+as Nix code. You can define derivations that run Python or Julia code,
+and objects can be exchanged between R and Python by using `rxp_py2r()`
+and `rxp_r2py()`, or by serializing to a common format such as JSON.
+By default, calling `rxp_populate()` also builds the pipeline, but it’s
+possible to only generate the `pipeline.nix` file and build the pipeline
+later using:
 
 ``` r
 rxp_make()
 ```
 
-The build process assumes the presence of a `default.nix` file which
-defines the computational environment the pipeline runs in; this
-`default.nix` file can be generated by using the `{rix}` package. This
-`default.nix` defines an environment with R, R packages (and also Python
-and/or Julia and their packages if needed), Quarto, and all required
-system-level dependencies pinned to a specific date to ensure
-reproducibility.
+The build process assumes the presence of a `default.nix` file that
+defines the computational environment the pipeline runs in; this file
+can be generated with the {rix} package. The `default.nix` typically
+defines an environment with R and required R packages (and optionally
+Python/Julia and their packages), Quarto, and any necessary system-level
+dependencies pinned to a specific date to ensure reproducibility.
 
[…]
