+++ README.md
@@ -6,30 +6,29 @@ v2](https://github.com/b-rodrigues/rixpress/actions/workflows/rhub.yaml/badge.sv
 [![Status at rOpenSci Software Peer
 Review](https://badges.ropensci.org/706_status.svg)](https://github.com/ropensci/software-review/issues/706)
 
-<p>
-
 If you want to watch a 2-Minute video introduction, click the image
 below:
-</p>
 
 <a href="https://www.youtube.com/watch?v=a1eNG9TFZ_o" target="_blank" rel="noopener noreferrer">
 <img src="https://raw.githubusercontent.com/b-rodrigues/rixpress/refs/heads/main/video_thumbnail.png" alt="Video Thumbnail" style="width:100%; max-width:560px; height:auto; display:block; margin:0 auto;">
 </a>
 
-`{rixpress}` provides a framework for building multilanguage reproducible
-analytical pipelines by leveraging Nix’s build automation capabilities.
-One of the design goals of `{rixpress}` is to mimic the user experience of
-the `{targets}` package, and it is heavily inspired by that workflow. It
-builds on the `{rix}` package, which provides helper functions to define
-reproducible development environments as code using Nix, ensuring the
-pipeline runs in a fully reproducible Nix-managed environment. `{rixpress}`
-only requires users to write the pipeline using familiar R code.
+`{rixpress}` provides a framework for building multilanguage
+reproducible analytical pipelines by leveraging Nix’s build automation
+capabilities. One of the design goals of `{rixpress}` is to mimic the
+user experience of the `{targets}` package, and it is heavily inspired
+by that workflow. It builds on the `{rix}` package, which provides
+helper functions to define reproducible development environments as code
+using Nix, ensuring the pipeline runs in a fully reproducible
+Nix-managed environment. `{rixpress}` only requires users to write the
+pipeline using familiar R code.
 
 `rixpress` focuses on “micropipelines”: pipelines executed on a single
 machine for small-to-medium sized projects.
 
 For example, this R script defines a list of *derivations* defined by
-functions prefixed with `rxp_*()`, which is then passed to `rxp_populate()`:
+functions prefixed with `rxp_*()`, which is then passed to
+`rxp_populate()`:
 
 ``` r
 library(rixpress)
@@ -73,8 +72,8 @@ Running `rxp_populate()` generates a `pipeline.nix` file, which contains
 the build instructions for all derivations and final outputs expressed
 as Nix code. You can define derivations that run Python or Julia code,
 and objects can be exchanged between R and Python by using `rxp_py2r()`
-and `rxp_r2py()`, or by serializing to a common format such as JSON.
-By default, calling `rxp_populate()` also builds the pipeline, but it’s
+and `rxp_r2py()`, or by serializing to a common format such as JSON. By
+default, calling `rxp_populate()` also builds the pipeline, but it’s
 possible to only generate the `pipeline.nix` file and build the pipeline
 later using:
 
@@ -90,7 +89,7 @@ Python/Julia and their packages), Quarto, and any necessary system-level
 dependencies pinned to a specific date to ensure reproducibility.
 
 In the example above, the first derivation reads `mtcars.csv` (in the
-example it’s pipe-separated, i.e. a `.psv` file). Each output (for
+example it’s pipe-separated, i.e. a `.psv` file). Each output (for
 example, `mtcars`, `mtcars_am`, `mtcars_head`, `mtcars_tail`,
 `mtcars_mpg`, `page`) is built by Nix within the environment defined by
 `default.nix`. Concretely, {rix} makes using Nix as a package manager
@@ -117,15 +116,15 @@ DAG
 
 Because the pipeline is built using Nix, outputs are stored in the Nix
 store under `/nix/store/`. To make working with these outputs easier,
[…]
