+++ README.Rmd
@@ -13,31 +13,34 @@ knitr::opts_chunk$set(
 
 # rixpress: Reproducible Analytical Pipelines with `Nix`
 
-[![R-hub v2](https://github.com/b-rodrigues/rixpress/actions/workflows/rhub.yaml/badge.svg)](https://github.com/b-rodrigues/rixpress/actions/workflows/rhub.yaml/)
-[![Status at rOpenSci Software Peer Review](https://badges.ropensci.org/706_status.svg)](https://github.com/ropensci/software-review/issues/706)
+[![R-hub
+v2](https://github.com/b-rodrigues/rixpress/actions/workflows/rhub.yaml/badge.svg)](https://github.com/b-rodrigues/rixpress/actions/workflows/rhub.yaml/)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/706_status.svg)](https://github.com/ropensci/software-review/issues/706)
 
-<p>If you want to watch a 2-Minute video introduction, click the image below:</p>
+If you want to watch a 2-Minute video introduction, click the image
+below:
 
 <a href="https://www.youtube.com/watch?v=a1eNG9TFZ_o" target="_blank" rel="noopener noreferrer">
-    <img src="https://raw.githubusercontent.com/b-rodrigues/rixpress/refs/heads/main/video_thumbnail.png" alt="Video Thumbnail" style="width:100%; max-width:560px; height:auto; display:block; margin:0 auto;">
+<img src="https://raw.githubusercontent.com/b-rodrigues/rixpress/refs/heads/main/video_thumbnail.png" alt="Video Thumbnail" style="width:100%; max-width:560px; height:auto; display:block; margin:0 auto;">
 </a>
 
-This R package provides a framework for building multilanguage reproducible
-analytical pipelines by leveraging `Nix`'s build automation capabilities. One of
-the design goals of `{rixpress}` is to mimic the user experience of the
-`{targets}` package, and is thus heavily inspired by it. It builds upon the
-`{rix}` package, which provides helper function to define reproducible
-development environments as code using `Nix`, ensuring the pipeline runs in a
-fully reproducible `Nix`-managed environment. `{rixpress}` only requires users
-to write the pipeline using familiar R code.
+`{rixpress}` provides a framework for building multilanguage reproducible
+analytical pipelines by leveraging Nix’s build automation capabilities.
+One of the design goals of `{rixpress}` is to mimic the user experience of
+the `{targets}` package, and it is heavily inspired by that workflow. It
+builds on the `{rix}` package, which provides helper functions to define
+reproducible development environments as code using Nix, ensuring the
+pipeline runs in a fully reproducible Nix-managed environment. `{rixpress}`
+only requires users to write the pipeline using familiar R code.
 
-`rixpress` focuses on “micropipelines”: pipelines executed on a single machine
-for small-to-medium projects.
+`rixpress` focuses on “micropipelines”: pipelines executed on a single
+machine for small-to-medium sized projects.
 
-For example, this R script defines a list of *derivations* defined by functions
-prefixed with `rxp_*()`, which is then passed to `rxp_populate()`:
+For example, this R script defines a list of *derivations* defined by
+functions prefixed with `rxp_*()`, which is then passed to `rxp_populate()`:
 
-``` r
+```r
 library(rixpress)
 
 list(
@@ -75,104 +78,78 @@ list(
   rxp_populate()
 ```
 
-Running `rxp_populate()` generates a `pipeline.nix` file, which contains all the
-build instructions of all the derivations and final outputs as `Nix` code. It is
-possible to define derivations that run Python or Julia code as well, and
-objects can be passed to and from R or Python by using `rxp_py2r()` and
-`ryp_r2py()`, or by using a common serialization format such as JSON. By
[…]
