+++ R/rxp_lineage.R
@@ -1,0 +1,255 @@
+#' Compute lineage for a derivation using _rixpress/dag.json
+#'
+#' rxp_lineage reads the DAG produced by rxp_populate (stored by default in
+#' _rixpress/dag.json at the project root) and returns a list with dependencies
+#' and reverse dependencies (derivations that depend on the target).
+#'
+#' By default the function expects the dag file at "_rixpress/dag.json". If your
+#' project uses a different path for the dag file, provide it via the dag_file
+#' argument.
+#'
+#' Transitive dependencies/reverse dependencies are marked with a "*" suffix so
+#' they can be visually distinguished from immediate neighbors.
+#'
+#' @param name Character scalar. Name of the derivation to inspect.
+#' @param dag_file Character scalar. Path to dag.json. Defaults to "_rixpress/dag.json".
+#' @param transitive Logical, default TRUE. If TRUE, return transitive closure and
+#'   mark transitive-only nodes with "*". If FALSE, returns only immediate neighbors.
+#' @param include_self Logical, default FALSE. If TRUE, include `name` in results.
+#' @return A list with elements:
+#'   - dependencies: character vector of derivation names this derivation depends on
+#'     (transitive-only entries have a trailing "*")
+#'   - reverse_dependencies: character vector of derivations that depend on this derivation
+#'     (transitive-only entries have a trailing "*")
+#' @family utilities
+#' @examples
+#' \dontrun{
+#' rxp_lineage("mtcars_mpg")
+#' rxp_lineage("mtcars_head", transitive = FALSE)
+#' rxp_lineage("page", dag_file = "/path/to/project/_rixpress/dag.json")
+#' }
+#' @export
+rxp_lineage <- function(
+  name,
+  dag_file = file.path("_rixpress", "dag.json"),
+  transitive = TRUE,
+  include_self = FALSE
+) {
+  if (missing(name) || !is.character(name) || length(name) != 1) {
+    stop("`name` must be a single character string with the derivation name.")
+  }
+  if (!is.character(dag_file) || length(dag_file) != 1) {
+    stop("`dag_file` must be a single character string path to dag.json.")
+  }
+
+  if (!file.exists(dag_file)) {
+    stop(
+      "Could not find dag file at: ",
+      dag_file,
+      ". By default rxp_lineage expects '_rixpress/dag.json'. ",
+      "If your dag.json is elsewhere, pass dag_file explicitly."
+    )
+  }
+
+  if (!requireNamespace("jsonlite", quietly = TRUE)) {
+    stop(
+      "Package 'jsonlite' is required to read dag.json. Install it and retry."
+    )
+  }
+
+  dag <- jsonlite::fromJSON(dag_file, simplifyVector = FALSE)
+
+  if (
+    !is.list(dag) || is.null(dag$derivations) || length(dag$derivations) == 0
+  ) {
+    stop("Invalid dag.json: no derivations found.")
+  }
+
+  derivs <- dag$derivations
+
+  # Helper to extract the string name from derivation$deriv_name
+  extract_name <- function(d) {
+    dn <- d$deriv_name
+    if (is.null(dn)) {
+      return(NA_character_)
+    }
+    # deriv_name in the JSON might be an array/list; take the first non-null element
+    if (is.list(dn) || length(dn) > 1) {
+      for (el in dn) {
+        if (!is.null(el) && nzchar(as.character(el))) return(as.character(el))
+      }
+      return(NA_character_)
+    }
+    as.character(dn)
+  }
+
+  all_names <- vapply(derivs, extract_name, character(1))
[…]
