+++ R/rxp_trace.R
@@ -1,0 +1,249 @@
+#' Trace lineage for derivations using _rixpress/dag.json
+#'
+#' Reads the DAG produced by rxp_populate (stored by default in
+#' _rixpress/dag.json) and prints dependencies and reverse dependencies.
+#'
+#' @param name Optional character scalar. Name of the derivation to inspect.
+#'   If NULL (default), the function prints the whole pipeline (inverted global view).
+#' @param dag_file Character scalar. Path to dag.json. Defaults to "_rixpress/dag.json".
+#' @param transitive Logical, default TRUE. If TRUE, show transitive closure and
+#'   mark transitive-only nodes with "*". If FALSE, show immediate neighbours only.
+#' @param include_self Logical, default FALSE. If TRUE, include `name` itself in the results.
+#' @return Invisibly, a named list mapping each inspected derivation name to a
+#'   list with elements:
+#'     - dependencies
+#'     - reverse_dependencies
+#'   The function also prints a tree representation to the console.
+#' @export
+rxp_trace <- function(
+  name = NULL,
+  dag_file = file.path("_rixpress", "dag.json"),
+  transitive = TRUE,
+  include_self = FALSE
+) {
+  if (!file.exists(dag_file)) {
+    stop(
+      "Could not find dag file at: ",
+      dag_file,
+      ". By default rxp_trace expects '_rixpress/dag.json'. ",
+      "If your dag.json is elsewhere, pass dag_file explicitly."
+    )
+  }
+  if (!requireNamespace("jsonlite", quietly = TRUE)) {
+    stop(
+      "Package 'jsonlite' is required to read dag.json. Install it and retry."
+    )
+  }
+
+  load_dag <- function(path) {
+    dag <- jsonlite::fromJSON(path, simplifyVector = FALSE)
+    if (
+      !is.list(dag) || is.null(dag$derivations) || length(dag$derivations) == 0
+    ) {
+      stop("Invalid dag.json: no derivations found.")
+    }
+    dag$derivations
+  }
+  derivs <- load_dag(dag_file)
+
+  extract_name <- function(d) {
+    dn <- d$deriv_name
+    if (is.null(dn)) {
+      return(NA_character_)
+    }
+    if (is.list(dn) || length(dn) > 1) {
+      for (el in dn) {
+        if (!is.null(el) && nzchar(as.character(el))) return(as.character(el))
+      }
+      return(NA_character_)
+    }
+    as.character(dn)
+  }
+
+  all_names <- vapply(derivs, extract_name, character(1))
+  if (anyNA(all_names)) {
+    stop("Found derivations with missing or unparsable names in dag.json.")
+  }
+
+  if (!is.null(name) && !name %in% all_names) {
+    stop(
+      "Derivation '",
+      name,
+      "' not found in dag.json (available: ",
+      paste(head(all_names, 20), collapse = ", "),
+      if (length(all_names) > 20) ", ..." else "",
+      ")."
+    )
+  }
+
+  make_depends_map <- function(derivs, names) {
+    out <- setNames(vector("list", length(names)), names)
+    for (i in seq_along(derivs)) {
+      deps <- derivs[[i]]$depends
+      if (is.null(deps)) {
+        deps <- character(0)
+      } else {
+        deps <- as.character(unlist(deps, use.names = FALSE))
+        deps <- deps[nzchar(deps)]
+      }
+      # only keep DAG-internal deps, and drop self-loops
+      deps <- deps[deps %in% names & deps != names[i]]
+      out[[names[i]]] <- unique(deps)
+    }
+    out
+  }
+
+  depends_map <- make_depends_map(derivs, all_names)
[…]
