+++ man/rxp_trace.Rd
@@ -1,0 +1,35 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rxp_trace.R
+\name{rxp_trace}
+\alias{rxp_trace}
+\title{Trace lineage for derivations using _rixpress/dag.json}
+\usage{
+rxp_trace(
+  name = NULL,
+  dag_file = file.path("_rixpress", "dag.json"),
+  transitive = TRUE,
+  include_self = FALSE
+)
+}
+\arguments{
+\item{name}{Optional character scalar. Name of the derivation to inspect.
+If NULL (default), the function prints the whole pipeline (inverted global view).}
+
+\item{dag_file}{Character scalar. Path to dag.json. Defaults to "_rixpress/dag.json".}
+
+\item{transitive}{Logical, default TRUE. If TRUE, show transitive closure and
+mark transitive-only nodes with "*". If FALSE, show immediate neighbours only.}
+
+\item{include_self}{Logical, default FALSE. If TRUE, include \code{name} itself in the results.}
+}
+\value{
+Invisibly, a named list mapping each inspected derivation name to a
+list with elements:
+- dependencies
+- reverse_dependencies
+The function also prints a tree representation to the console.
+}
+\description{
+Reads the DAG produced by rxp_populate (stored by default in
+_rixpress/dag.json) and prints dependencies and reverse dependencies.
+}
