+++ NAMESPACE
@@ -26,6 +26,7 @@ export(rxp_r2py)
 export(rxp_r_file)
 export(rxp_read)
 export(rxp_rmd)
+export(rxp_trace)
 export(rxp_visnetwork)
 export(rxp_write_dag)
 importFrom(igraph,V)
