+++ man/rxp_load.Rd
@@ -53,6 +53,7 @@ Other utilities:
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
-\code{\link{rxp_read}()}
+\code{\link{rxp_read}()},
+\code{\link{rxp_trace}()}
 }
 \concept{utilities}
