+++ man/rxp_read.Rd
@@ -49,6 +49,7 @@ Other utilities:
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
-\code{\link{rxp_load}()}
+\code{\link{rxp_load}()},
+\code{\link{rxp_trace}()}
 }
 \concept{utilities}
