+++ man/rxp_inspect.Rd
@@ -50,6 +50,7 @@ Other utilities:
 \code{\link{rxp_init}()},
 \code{\link{rxp_list_logs}()},
 \code{\link{rxp_load}()},
-\code{\link{rxp_read}()}
+\code{\link{rxp_read}()},
+\code{\link{rxp_trace}()}
 }
 \concept{utilities}
