+++ man/rxp_list_logs.Rd
@@ -33,6 +33,7 @@ Other utilities:
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_load}()},
-\code{\link{rxp_read}()}
+\code{\link{rxp_read}()},
+\code{\link{rxp_trace}()}
 }
 \concept{utilities}
