+++ vignettes/core-functions.Rmd
@@ -80,7 +80,7 @@ dependencies between derivations are defined.
 
 ## Generating the pipeline
 
-Let's stop here and build our pipeline. First, we need to define a list of
+Let's stop here and generate our pipeline. First, we need to define a list of
 derivations:
 
 ```{r, eval = FALSE}
@@ -225,54 +225,9 @@ I recommend always using two separate scripts:
 You can quickly create these scripts using the `rxp_init()` function, which
 generates both files with starter code to help you get started quickly.
 
-## Building and inspecting outputs
+## Optional steps before building the pipeline
 
-When you run `gen-pipeline.R` (or execute its contents line-by-line), the
-environment defined in `default.nix` is used. (It's also possible to define
-separate environments for different derivations, which we'll cover in a later
-vignette.) Upon successful execution, you will see:
-
-```
-Build process started...
-
-
-Build successful! Run `rxp_inspect()` for a summary.
-Read individual derivations using `rxp_read()` or
-load them into the global environment using `rxp_load()`.
-```
-
-Now you can follow these instructions:
-
-1. Use `rxp_inspect()` to see where the outputs are located. This function is
-   particularly useful if the pipeline fails, as it shows which derivations
-   succeeded and which failed.
-2. Use `rxp_read("mtcars_mpg")` to read the object into your current R session,
-   or `rxp_load("mtcars_mpg")` to load it directly into your global environment.
-3. Alternatively, use `rxp_copy("mtcars_mpg")` to create a folder called
-   `pipeline-outputs` containing `mtcars_mpg` as an `.rds` file. If you call
-   `rxp_copy()` without arguments, all pipeline outputs will be copied to this
-   folder.
-
-## No-op builds for individual derivations
-
-You can disable building a specific derivation by setting its `noop_build`
-parameter to `TRUE`. This creates a no-op build, a placeholder derivation that
-performs no work:
-
-```r
-rxp_r(
-  name = turtles,
-  expr = occurrence(species, geometry = atlantic),
-  noop_build = TRUE
-)
-```
-
-Any derivations that depend on a no-op build will themselves also resolve to
-no-op builds. This can be useful when prototyping or debugging a pipeline,
-allowing you to skip expensive or unnecessary computations while keeping the
-dependency graph intact.
-
-## DAG Representation of the Pipeline
+### Graphical representation of the pipeline's DAG
 
 It's often helpful to visualise your pipeline as a DAG (directed acyclic graph).
 By default, the `build` argument of `rxp_populate()` is `FALSE`, so calling
@@ -326,6 +281,126 @@ which saves the `dag.dot` object in the project's `_rixpress/` folder.
 After reviewing the DAG, you can build the pipeline by running `rxp_make()`
 instead of modifying your original `rxp_populate()` call.
 
+### Tracing the lineage of derivations
+
+It is possible to also trace the lineage of individual derivations using
+`rxp_trace()`. For example:
+
+```r
+rxp_trace("mtcars_mpg")
+```
+
+will return:
+
+```r
+==== Lineage for: mtcars_mpg ====
+Dependencies (ancestors):
+  - mtcars_head
+    - mtcars_am*
+      - mtcars*
+  - mtcars_tail
+    - mtcars_head*
+
+Reverse dependencies (children):
+          - page
+
+Note: '*' marks transitive dependencies (depth >= 2).
+```
+
+This makes it quite easy to quickly double check whether derivations
+were defined correctly. A `*` symbol next to a derivation's name indicates
+it is a transitive dependency. Calling `rxp_trace()` without arguments
+shows the whole graph:
+
+```r
+rxp_trace()
+```
+
+```r
[…]
