+++ NAMESPACE
@@ -40,3 +40,4 @@ importFrom(processx,run)
 importFrom(stats,setNames)
 importFrom(utils,capture.output)
 importFrom(utils,globalVariables)
+importFrom(utils,head)
