### 2025-09-01 1a28ed5
Add rxp_gc function for Nix garbage collection
