+++ R/rxp_gc.R
@@ -1,0 +1,238 @@
+#' Clean up rixpress build artifacts with Nix garbage collection
+#'
+#' Safely removes unused build artifacts from a rixpress project by invoking
+#' Nix garbage collection (`nix-store --gc`). Supports two modes:
+#'
+#' * **Default mode (`keep_since = NULL`)**: Run full Nix garbage collection
+#'   to delete all unreferenced store paths.
+#' * **Date-based retention (`keep_since` supplied)**: Keep artifacts from builds
+#'   newer than the given date, delete artifacts from older builds, and protect
+#'   recent builds with temporary GC roots.
+#'
+#' The function is **destructive**: deleted store paths cannot be recovered.
+#' Always test with `dry_run = TRUE` first.
+#'
+#' @param keep_since `NULL` (default) or a `Date`/string in `"YYYY-MM-DD"` format.
+#'   - `NULL`: run full Nix GC.
+#'   - Date/string: retain builds newer than this date, delete older builds.
+#' @param project_path Path to the project root (default `"."`).
+#'   The project must contain an `_rixpress/` directory with build logs.
+#' @param dry_run Logical (default `TRUE`). If `TRUE`, show what would be done
+#'   without deleting anything or running GC.
+#' @param timeout_sec Numeric (default `300`). Timeout (in seconds) for
+#'   `nix-store` commands.
+#' @param verbose Logical (default `FALSE`). If `TRUE`, print full output of
+#'   `nix-store --gc`. Otherwise, show only summary lines.
+#'
+#' @return Invisibly returns a list with:
+#'   - `kept`: filenames of logs kept
+#'   - `deleted`: filenames of logs scheduled for deletion
+#'   - `protected`: number of GC roots created for recent builds
+#'
+#' @details
+#' ## Concurrency
+#' Uses a lock file in `tempdir()` to prevent concurrent runs. If a process dies
+#' and leaves a stale lock, it will be removed if older than `timeout_sec`.
+#'
+#' ## Safety mechanisms
+#' - All destructive operations require confirmation (`askYesNo()`).
+#' - Invalid or missing store paths are skipped with a warning.
+#' - Recent builds are protected with temporary GC roots.
+#' - Use `dry_run = TRUE` to preview deletions.
+#'
+#' ## Platform notes
+#' - Requires Nix installed and accessible in `PATH`.
+#' - Tested on Unix-like systems. On Windows, PID detection for concurrency may
+#'   be less reliable due to faster PID recycling.
+#'
+#' @examples
+#' \dontrun{
+#' # Dry run of full garbage collection
+#' rxp_gc()
+#'
+#' # Dry run of date-based retention
+#' rxp_gc(keep_since = "2025-01-01")
+#'
+#' # Actually delete all unreferenced store paths
+#' rxp_gc(dry_run = FALSE)
+#' }
+#'
+#' @export
+rxp_gc <- function(
+  keep_since = NULL,
+  project_path = ".",
+  dry_run = TRUE,
+  timeout_sec = 300,
+  verbose = FALSE
+) {
+  # --- Concurrency Check ---
+  lock_file <- file.path(tempdir(), "rixpress_gc.lock")
+  if (file.exists(lock_file)) {
+    tryCatch({
+      lock_info <- readLines(lock_file)
+      lock_pid <- as.numeric(lock_info[1])
+      lock_time <- as.POSIXct(lock_info[2])
+      
+      alive <- FALSE
+      if (.Platform$OS.type == "unix" && !is.na(lock_pid)) {
+        res <- system2("ps", c("-p", lock_pid), stdout = TRUE, stderr = FALSE)
+        alive <- any(grepl(lock_pid, res))
+      }
+      
+      if (alive) {
[…]
