+++ R/rxp_gc.R
@@ -68,30 +68,39 @@ rxp_gc <- function(
   # --- Concurrency Check ---
   lock_file <- file.path(tempdir(), "rixpress_gc.lock")
   if (file.exists(lock_file)) {
-    tryCatch({
-      lock_info <- readLines(lock_file)
-      lock_pid <- as.numeric(lock_info[1])
-      lock_time <- as.POSIXct(lock_info[2])
-      
-      alive <- FALSE
-      if (.Platform$OS.type == "unix" && !is.na(lock_pid)) {
-        res <- system2("ps", c("-p", lock_pid), stdout = TRUE, stderr = FALSE)
-        alive <- any(grepl(lock_pid, res))
-      }
-      
-      if (alive) {
-        stop("Another rxp_gc process appears to be running (PID: ", lock_pid, 
-             " started at ", lock_time, "). If this is incorrect, remove: ", lock_file)
-      }
-      
-      if (difftime(Sys.time(), lock_time, units = "secs") > timeout_sec) {
-        file.remove(lock_file)
-        message("Removed stale lock file from ", lock_time)
-      } else if (!alive) {
-        message("Stale lock file found (PID dead). Removing.")
-        file.remove(lock_file)
-      }
-    }, error = function(e) file.remove(lock_file))
+    tryCatch(
+      {
+        lock_info <- readLines(lock_file)
+        lock_pid <- as.numeric(lock_info[1])
+        lock_time <- as.POSIXct(lock_info[2])
+
+        alive <- FALSE
+        if (.Platform$OS.type == "unix" && !is.na(lock_pid)) {
+          res <- system2("ps", c("-p", lock_pid), stdout = TRUE, stderr = FALSE)
+          alive <- any(grepl(lock_pid, res))
+        }
+
+        if (alive) {
+          stop(
+            "Another rxp_gc process appears to be running (PID: ",
+            lock_pid,
+            " started at ",
+            lock_time,
+            "). If this is incorrect, remove: ",
+            lock_file
+          )
+        }
+
+        if (difftime(Sys.time(), lock_time, units = "secs") > timeout_sec) {
+          file.remove(lock_file)
+          message("Removed stale lock file from ", lock_time)
+        } else if (!alive) {
+          message("Stale lock file found (PID dead). Removing.")
+          file.remove(lock_file)
+        }
+      },
+      error = function(e) file.remove(lock_file)
+    )
   }
   writeLines(c(as.character(Sys.getpid()), as.character(Sys.time())), lock_file)
   on.exit(if (file.exists(lock_file)) file.remove(lock_file), add = TRUE)
@@ -101,18 +110,27 @@ rxp_gc <- function(
     if (inherits(keep_since, "Date")) {
       # ok
     } else if (is.character(keep_since)) {
-      keep_since <- tryCatch(as.Date(keep_since),
-                             error = function(e) stop("Invalid 'keep_since'. Use 'YYYY-MM-DD'."))
+      keep_since <- tryCatch(
+        as.Date(keep_since),
+        error = function(e) stop("Invalid 'keep_since'. Use 'YYYY-MM-DD'.")
+      )
     } else stop("'keep_since' must be a Date or a 'YYYY-MM-DD' string.")
   }
-  if (!is.numeric(timeout_sec) || timeout_sec <= 0) stop("'timeout_sec' must be positive.")
+  if (!is.numeric(timeout_sec) || timeout_sec <= 0)
+    stop("'timeout_sec' must be positive.")
 
   nix_bin <- Sys.which("nix-store")
-  if (nix_bin == "") stop("nix-store not found. Ensure Nix is installed and in PATH.")
+  if (nix_bin == "")
+    stop("nix-store not found. Ensure Nix is installed and in PATH.")
 
   all_logs <- rxp_list_logs(project_path)
[…]
