### 2025-09-01 40d8206
Update rxp_gc.R
