+++ R/rxp_gc.R
@@ -17,7 +17,7 @@
 #'   - Date/string: retain builds newer than this date, delete older builds.
 #' @param project_path Path to the project root (default `"."`).
 #'   The project must contain an `_rixpress/` directory with build logs.
-#' @param dry_run Logical (default `TRUE`). If `TRUE`, show what would be done
+#' @param dry_run Logical (default `FALSE`). If `TRUE`, show what would be done
 #'   without deleting anything or running GC.
 #' @param timeout_sec Numeric (default `300`). Timeout (in seconds) for
 #'   `nix-store` commands.
@@ -48,261 +48,152 @@
 #' @examples
 #' \dontrun{
 #' # Dry run of full garbage collection
-#' rxp_gc()
+#' rxp_gc(dry_run = TRUE)
 #'
 #' # Dry run of date-based retention
-#' rxp_gc(keep_since = "2025-01-01")
+#' rxp_gc(keep_since = "2025-01-01", dry_run = TRUE)
 #'
 #' # Actually delete all unreferenced store paths
-#' rxp_gc(dry_run = FALSE)
+#' rxp_gc()
 #' }
 #'
 #' @export
-rxp_gc <- function(
-  keep_since = NULL,
-  project_path = ".",
-  dry_run = TRUE,
-  timeout_sec = 300,
-  verbose = FALSE
-) {
-  # --- Concurrency Check ---
-  lock_file <- file.path(tempdir(), "rixpress_gc.lock")
-  if (file.exists(lock_file)) {
-    tryCatch(
-      {
-        lock_info <- readLines(lock_file)
-        lock_pid <- as.numeric(lock_info[1])
-        lock_time <- as.POSIXct(lock_info[2])
+rxp_gc <- function(keep_since = NULL,
+                   project_path = ".",
+                   dry_run = FALSE,
+                   timeout_sec = 300,
+                   verbose = FALSE) {
 
-        alive <- FALSE
-        if (.Platform$OS.type == "unix" && !is.na(lock_pid)) {
-          res <- system2("ps", c("-p", lock_pid), stdout = TRUE, stderr = FALSE)
-          alive <- any(grepl(lock_pid, res))
-        }
-
-        if (alive) {
-          stop(
-            "Another rxp_gc process appears to be running (PID: ",
-            lock_pid,
-            " started at ",
-            lock_time,
-            "). If this is incorrect, remove: ",
-            lock_file
-          )
-        }
-
-        if (difftime(Sys.time(), lock_time, units = "secs") > timeout_sec) {
-          file.remove(lock_file)
-          message("Removed stale lock file from ", lock_time)
-        } else if (!alive) {
-          message("Stale lock file found (PID dead). Removing.")
-          file.remove(lock_file)
-        }
-      },
-      error = function(e) file.remove(lock_file)
-    )
-  }
-  writeLines(c(as.character(Sys.getpid()), as.character(Sys.time())), lock_file)
-  on.exit(if (file.exists(lock_file)) file.remove(lock_file), add = TRUE)
-
-  # --- Input Validation ---
-  if (!is.null(keep_since)) {
-    if (inherits(keep_since, "Date")) {
-      # ok
-    } else if (is.character(keep_since)) {
-      keep_since <- tryCatch(
-        as.Date(keep_since),
-        error = function(e) stop("Invalid 'keep_since'. Use 'YYYY-MM-DD'.")
-      )
-    } else stop("'keep_since' must be a Date or a 'YYYY-MM-DD' string.")
-  }
-  if (!is.numeric(timeout_sec) || timeout_sec <= 0)
-    stop("'timeout_sec' must be positive.")
-
-  nix_bin <- Sys.which("nix-store")
-  if (nix_bin == "")
[…]
