+++ R/rxp_gc.R
@@ -1,62 +1,91 @@
-#' Clean up rixpress build artifacts with Nix garbage collection
+#' Garbage Collect Rixpress Build Artifacts and Logs
 #'
-#' Safely removes unused build artifacts from a rixpress project by invoking
-#' Nix garbage collection (`nix-store --gc`). Supports two modes:
+#' This function performs garbage collection on Nix store paths and build log files
+#' generated by rixpress. It can operate in two modes: full garbage collection
+#' (when \code{keep_since = NULL}) or targeted deletion based on log file age.
 #'
-#' * **Default mode (`keep_since = NULL`)**: Run full Nix garbage collection
-#'   to delete all unreferenced store paths.
-#' * **Date-based retention (`keep_since` supplied)**: Keep artifacts from builds
-#'   newer than the given date, delete artifacts from older builds, and protect
-#'   recent builds with temporary GC roots.
+#' @param keep_since Date or character string (YYYY-MM-DD format). If provided,
+#'   only build logs older than this date will be targeted for deletion, along
+#'   with their associated Nix store paths. If \code{NULL}, performs a full
+#'   Nix garbage collection. Default is \code{NULL}.
+#' @param project_path Character string specifying the path to the project
+#'   directory containing the \code{_rixpress} folder with build logs.
+#'   Default is \code{"."} (current directory).
+#' @param dry_run Logical. If \code{TRUE}, shows what would be deleted without
+#'   actually performing any deletions. Useful for previewing the cleanup
+#'   operation. Default is \code{TRUE}.
+#' @param timeout_sec Numeric. Timeout in seconds for individual Nix commands.
+#'   Also used for concurrency lock expiration. Default is 300 seconds.
+#' @param verbose Logical. If \code{TRUE}, provides detailed output including
+#'   full paths, command outputs, and diagnostic information about references
+#'   preventing deletion. Default is \code{FALSE}.
 #'
-#' The function is **destructive**: deleted store paths cannot be recovered.
-#' Always test with `dry_run = TRUE` first.
-#'
-#' @param keep_since `NULL` (default) or a `Date`/string in `"YYYY-MM-DD"` format.
-#'   - `NULL`: run full Nix GC.
-#'   - Date/string: retain builds newer than this date, delete older builds.
-#' @param project_path Path to the project root (default `"."`).
-#'   The project must contain an `_rixpress/` directory with build logs.
-#' @param dry_run Logical (default `FALSE`). If `TRUE`, show what would be done
-#'   without deleting anything or running GC.
-#' @param timeout_sec Numeric (default `300`). Timeout (in seconds) for
-#'   `nix-store` commands.
-#' @param verbose Logical (default `FALSE`). If `TRUE`, print full output of
-#'   `nix-store --gc`. Otherwise, show only summary lines.
-#'
-#' @return Invisibly returns a list with:
-#'   - `kept`: filenames of logs kept
-#'   - `deleted`: filenames of logs scheduled for deletion
-#'   - `protected`: number of GC roots created for recent builds
+#' @return Invisibly returns a list with cleanup summary information:
+#' \itemize{
+#'   \item \code{kept}: Vector of build log filenames that were kept
+#'   \item \code{deleted}: Vector of build log filenames targeted for deletion
[…]
