+++ man/rxp_read.Rd
@@ -46,6 +46,7 @@ returned instead.
 Other utilities: 
 \code{\link{print.rxp_derivation}()},
 \code{\link{rxp_copy}()},
+\code{\link{rxp_gc}()},
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_list_logs}()},
