+++ man/rxp_list_logs.Rd
@@ -30,6 +30,7 @@ in the project's _rixpress directory.
 Other utilities: 
 \code{\link{print.rxp_derivation}()},
 \code{\link{rxp_copy}()},
+\code{\link{rxp_gc}()},
 \code{\link{rxp_init}()},
 \code{\link{rxp_inspect}()},
 \code{\link{rxp_load}()},
