+++ NAMESPACE
@@ -7,6 +7,7 @@ export(rxp_copy)
 export(rxp_dag_for_ci)
 export(rxp_export_artifacts)
 export(rxp_ga)
+export(rxp_gc)
 export(rxp_ggdag)
 export(rxp_import_artifacts)
 export(rxp_init)
