+++ pkgdown.yml
@@ -12,7 +12,7 @@ articles:
   polyglot: polyglot.html
   scope: scope.html
   tutorial: tutorial.html
-last_built: 2025-09-02T15:11Z
+last_built: 2025-09-03T14:40Z
 urls:
   reference: https://b-rodrigues.github.io/rixpress/reference
   article: https://b-rodrigues.github.io/rixpress/articles
