### 2025-09-03 18b4985
Add vignette for importing data files with rixpress
This vignette explains how to import data files using the {rixpress} package, covering local files, files from URLs, multiple files from a directory, and files with dependencies like shapefiles.
