+++ vignettes/read_files.Rmd
@@ -1,0 +1,161 @@
+---
+title: "Importing Data Files"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{importing-data}
+  %\VignetteEngine{knitr::rmmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+A crucial first step in any data analysis pipeline is importing data. 
+The `{rixpress}` package provides a flexible set of functions, `rxp_r_file`, `rxp_py_file`, and `rxp_jl_file`, to handle various data import scenarios in a reproducible way. 
+This vignette will guide you through the common use cases.
+
+For more examples, check out the [rixpress_demos repository](https://github.com/b-rodrigues/rixpress_demos/).
+
+## Importing a single local file
+
+The most straightforward case is reading a single data file from your local project directory. 
+You need to provide a `name` for the resulting R object, the `path` to the file, and a `read_function` to process it.
+
+```{r, eval = FALSE}
+library(rixpress)
+
+list(
+  rxp_r_file(
+    name = mtcars,
+    path = 'data/mtcars.csv',
+    read_function = \(x) (read.csv(file = x, sep = "|"))
+  ),
+...
+```
+
+In this example, `rxp_r_file` creates a derivation that:
+
+1. Copies `data/mtcars.csv` into a sandboxed build environment.
+2. Executes the provided anonymous function, `\(x) (read.csv(file = x, sep = "|"))`, where `x` is the path to the copied file inside the sandbox.
+3. Saves the resulting data frame as an object named `mtcars` for subsequent steps in the pipeline.
+
+## Importing a single file from the internet
+
+You can also directly import a file from a URL. Simply provide the URL as the `path`. 
+`{rixpress}` handles the download and ensures reproducibility by caching the file using its cryptographic hash.
+
+```{r, eval = FALSE}
+library(rixpress)
+
+list(
+  rxp_r_file(
+    name = mtcars,
+    path = 'https://raw.githubusercontent.com/b-rodrigues/rixpress_demos/refs/heads/master/basic_r/data/mtcars.csv',
+    read_function = \(x) (read.csv(file = x, sep = "|"))
+  ),
+...
+```
+
+Behind the scenes, `{rixpress}` uses Nix to fetch the file, ensuring that the exact same version of the file is used every time the pipeline is run.
+
+## Importing many files from a directory
+
+Often, you need to import and combine multiple files from a single directory. 
+  To do this, set the `path` argument to the directory's path. 
+Your `read_function` will then receive the path to this directory inside the build environment and must contain the logic to handle all the files within.
+
+Here is an example in R that reads all files in the `data` directory:
+
+```{r, eval = FALSE}
+library(rixpress)
+
+list(
+  rxp_r_file(
+    name = mtcars_r,
+    path = 'data',
+    read_function = \(x) {
+      (readr::read_delim(list.files(x, full.names = TRUE), delim = '|'))
+    }
+  )
+) |>
+  rxp_populate(project_path = ".")
+```
+
+And here's a similar example using Python, which calls a user-defined function `read_many_csvs` from an external script:
+
+```{r, eval = FALSE}
+library(rixpress)
+
+list(
+  rxp_py_file(
+    name = mtcars_py,
+    path = 'data',
+    read_function = "read_many_csvs",
+    user_functions = "functions.py"
+  )
+) |>
+  rxp_populate(project_path = ".")
+```
+
+Here is what the Python function looks like:
+
+```py
+import polars
+from pathlib import Path
+
+def read_many_csvs(dir_path):
+    folder = Path(dir_path)
+    csv_files = folder.glob("*.csv")
[…]
