### 2025-09-03 21ec7bb
Add 'importing-data' section to pkgdown config
