### 2025-09-03 3f8e6f6
switched from lambda to wrapper function
