+++ vignettes/importing-data.Rmd
@@ -14,16 +14,18 @@ knitr::opts_chunk$set(
 )
 ```
 
-A crucial first step in any data analysis pipeline is importing data. 
-The `{rixpress}` package provides a flexible set of functions, `rxp_r_file`, `rxp_py_file`, and `rxp_jl_file`, to handle various data import scenarios in a reproducible way. 
-This vignette will guide you through the common use cases.
+A crucial first step in any data analysis pipeline is importing data. The
+`{rixpress}` package provides a flexible set of functions, `rxp_r_file`,
+`rxp_py_file`, and `rxp_jl_file`, to handle various data import scenarios in a
+reproducible way. This vignette will guide you through the common use cases.
 
 For more examples, check out the [rixpress_demos repository](https://github.com/b-rodrigues/rixpress_demos/).
 
 ## Importing a single local file
 
-The most straightforward case is reading a single data file from your local project directory. 
-You need to provide a `name` for the resulting R object, the `path` to the file, and a `read_function` to process it.
+The most straightforward case is reading a single data file from your local
+project directory. You need to provide a `name` for the resulting R object, the
+`path` to the file, and a `read_function` to process it.
 
 ```{r, eval = FALSE}
 library(rixpress)
@@ -40,13 +42,15 @@ list(
 In this example, `rxp_r_file` creates a derivation that:
 
 1. Copies `data/mtcars.csv` into a sandboxed build environment.
-2. Executes the provided anonymous function, `\(x) (read.csv(file = x, sep = "|"))`, where `x` is the path to the copied file inside the sandbox.
+2. Executes the provided anonymous function, `\(x) (read.csv(file = x, sep =
+   "|"))`, where `x` is the path to the copied file inside the sandbox.
 3. Saves the resulting data frame as an object named `mtcars` for subsequent steps in the pipeline.
 
 ## Importing a single file from the internet
 
-You can also directly import a file from a URL. Simply provide the URL as the `path`. 
-`{rixpress}` handles the download and ensures reproducibility by caching the file using its cryptographic hash.
+You can also directly import a file from a URL. Simply provide the URL as the
+`path`. `{rixpress}` handles the download and ensures reproducibility by caching
+the file using its cryptographic hash.
 
 ```{r, eval = FALSE}
 library(rixpress)
@@ -60,13 +64,20 @@ list(
 ...
 ```
 
-Behind the scenes, `{rixpress}` uses Nix to fetch the file, ensuring that the exact same version of the file is used every time the pipeline is run.
+Behind the scenes, `{rixpress}` uses Nix to fetch the file, ensuring that the
+exact same version of the file is used every time the pipeline is run. This is
+the only time the build sandbox can access a remote file: it's because the file
+actually gets downloaded by Nix ahead of time. If you need to access data in
+real-time from an API, then you'll need to download the data yourself
+outside of `{rixpress}` pipeline, and then import it in the pipeline using
+`rxp_r_file()`.
 
 ## Importing many files from a directory
 
-Often, you need to import and combine multiple files from a single directory. 
-  To do this, set the `path` argument to the directory's path. 
-Your `read_function` will then receive the path to this directory inside the build environment and must contain the logic to handle all the files within.
+Often, you need to import and combine multiple files from a single directory. To
+do this, set the `path` argument to the directory's path. Your `read_function`
+will then receive the path to this directory inside the build environment and
+must contain the logic to handle all the files within.
 
 Here is an example in R that reads all files in the `data` directory:
 
@@ -85,7 +96,8 @@ list(
   rxp_populate(project_path = ".")
 ```
 
[…]
