+++ DESCRIPTION
@@ -1,15 +1,25 @@
 Package: rixpress
-Title: Build Reproducible Analytical Pipelines With Nix
+Title: Build Reproducible Analytical Pipelines with Nix
 Version: 0.2.0
-Authors@R:
-    person("Bruno", "Rodrigues", , "bruno@brodrigues.co", role = c("aut", "cre"))
+Authors@R: c(
+  person(given = "Bruno", family = "Rodrigues", email = "bruno@brodrigues.co", role = c("aut", "cre"),
+         comment = c(ORCID = "0000-0002-3211-3689")),
+  person(given = c("William", "Michael"), family = "Landau", email = "will.landau.oss@gmail.com", role = "rev",
+         comment = c("William reviewed the package (v. 0.2.0) for rOpenSci,
+                      see <https://github.com/ropensci/software-review/issues/706>")),
+  person(given = "Anthony", family = "Martinez", , "ajmartinez@usgs.gov", role = "rev",
+         comment = c(ORCID = "0000-0002-4295-0261", "Anthony reviewed the
+                     package (v. 0.2.0) for rOpenSci,
+                     see <https://github.com/ropensci/software-review/issues/625>"))
+  )
 Description: Streamlines the creation of reproducible analytical pipelines using
   `default.nix` expressions generated via `{rix}` for reproducibility. Define
-  derivations in R or Python, chain them into a composition of pure functions
-  and build the resulting pipeline using `Nix` as the underlying end-to-end build
-  tool. Functions to plot a DAG representation of the pipeline are included,
-  as well as functions to load and inspect intermediary results for interactive
-  analysis. User experience heavily inspired by the `{targets}` package.
+  derivations in R, Python or Julia, chain them into a composition of pure
+  functions and build the resulting pipeline using `Nix` as the underlying
+  end-to-end build tool. Functions to plot a DAG representation of the pipeline
+  are included, as well as functions to load and inspect intermediary results
+  for interactive analysis. User experience heavily inspired by the `{targets}`
+  package.
 License: GPL (>= 3)
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
