+++ vignettes/polyglot-julia.Rmd
@@ -41,7 +41,7 @@ configuration:
 library(rix)
 
 rix(
-  date = "2025-05-19",
+  date = "2025-09-04",
   r_pkgs = c(
     "arrow",
     "dplyr",
