### 2025-09-04 c8181ae
proper escaping for julia
