### 2025-09-05 c10854c
serialize_functions in rxp_file
