+++ R/rxp_io.R
@@ -196,6 +196,7 @@ sanitize_nix_env <- function(nix_env) {
 #' @param user_code Source/import statements for user functions
 #' @param out_name Name of the output object
 #' @param rel_path Relative path to input file
+#' @param serialize_str Language-specific serializer identifier or NULL for default
 #' @return Character vector of commands
 #' @keywords internal
 build_language_commands <- function(
@@ -203,7 +204,8 @@ build_language_commands <- function(
   read_func,
   user_code,
   out_name,
-  rel_path
+  rel_path,
+  serialize_str = NULL
 ) {
   user_lines <- if (nzchar(user_code)) {
     strsplit(user_code, "\n")[[1]]
@@ -213,28 +215,61 @@ build_language_commands <- function(
 
   switch(
     lang,
-    R = c(
-      "Rscript -e \"",
-      "source('libraries.R')",
-      user_lines,
-      sprintf("data <- do.call(%s, list('%s'))", read_func, rel_path),
-      sprintf("saveRDS(data, '%s')\"", out_name)
-    ),
-    Py = c(
-      "python -c \"",
-      "exec(open('libraries.py').read())",
-      user_lines,
-      sprintf("file_path = '%s'", rel_path),
-      sprintf("data = eval('%s')(file_path)", read_func),
-      sprintf("with open('%s', 'wb') as f:", out_name),
-      "    pickle.dump(data, f)",
-      "\""
-    ),
-    Jl = c(
-      "# Julia derivation build",
-      user_lines,
-      sprintf("data = %s(\"%s\")", read_func, rel_path)
-    ),
+    R = {
+      ser_fun <- if (is.null(serialize_str)) "saveRDS" else serialize_str
+      ser_call <- sprintf("%s(data, '%s')", ser_fun, out_name)
+      c(
+        "Rscript -e \"",
+        "source('libraries.R')",
+        user_lines,
+        sprintf("data <- do.call(%s, list('%s'))", read_func, rel_path),
+        paste0(ser_call, "\"")
+      )
+    },
+    Py = {
+      if (is.null(serialize_str)) {
+        ser_lines <- c(
+          sprintf("with open('%s', 'wb') as f:", out_name),
+          "    pickle.dump(data, f)"
+        )
+      } else {
+        ser_lines <- c(sprintf("%s(data, '%s')", serialize_str, out_name))
+      }
+      c(
+        "python -c \"",
+        "exec(open('libraries.py').read())",
+        user_lines,
+        sprintf("file_path = '%s'", rel_path),
+        sprintf("data = eval('%s')(file_path)", read_func),
+        ser_lines,
+        "\""
+      )
+    },
+    Jl = {
+      # Escape double quotes within -e string
+      user_lines_jl <- gsub("\"", "\\\\\"", user_lines, fixed = TRUE)
+      read_line <- sprintf("data = %s(\\\"%s\\\")", read_func, rel_path)
+      ser_line <- if (is.null(serialize_str)) {
+        paste0(
+          "using Serialization; ",
+          "io = open(\\\"",
+          out_name,
+          "\\\", \\\"w\\\"); ",
+          "serialize(io, data); ",
+          "close(io)"
+        )
+      } else {
+        sprintf("%s(data, \\\"%s\\\")", serialize_str, out_name)
+      }
+      c(
+        "julia -e \"",
+        "if isfile(\\\"libraries.jl\\\"); include(\\\"libraries.jl\\\"); end;",
+        user_lines_jl,
+        read_line,
+        ser_line,
+        "\""
+      )
+    },
     stop("Unsupported lang: ", lang)
   )
 }
@@ -249,6 +284,7 @@ build_language_commands <- function(
[…]
