### 2025-09-05 df8f087
serialize for rxp_file
