+++ man/process_serialize_function.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/rxp_io.R
+\name{process_serialize_function}
+\alias{process_serialize_function}
+\title{Process serialize function for different languages}
+\usage{
+process_serialize_function(
+  serialize_function,
+  lang,
+  parent_env = parent.frame()
+)
+}
+\arguments{
+\item{serialize_function}{Language-specific serializer (see above)}
+
+\item{lang}{Language string}
+
+\item{parent_env}{Environment for proper substitution (R only)}
+}
+\value{
+For R: function name string; For Py/Jl: NULL (use default) or function name string
+}
+\description{
+Mirrors the behaviour of rxp_r/rxp_py/rxp_jl:
+\itemize{
+\item R: supports bare symbol, character, or function; defaults to saveRDS
+\item Python: character function name, defaults to pickle.dump via with-open
+\item Julia: character function name, defaults to Serialization.serialize
+}
+}
+\keyword{internal}
