+++ man/create_rxp_derivation.Rd
@@ -10,7 +10,8 @@ create_rxp_derivation(
   lang,
   user_functions,
   nix_env,
-  env_var
+  env_var,
+  serialize_function = NULL
 )
 }
 \arguments{
@@ -25,6 +26,8 @@ create_rxp_derivation(
 \item{nix_env}{Character nix environment}
 
 \item{env_var}{Named list of environment variables}
+
+\item{serialize_function}{Character serializer identifier (for display)}
 }
 \value{
 rxp_derivation object
