+++ man/process_encoder.Rd
@@ -1,17 +1,13 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/rxp_io.R
-\name{process_serialize_function}
-\alias{process_serialize_function}
+\name{process_encoder}
+\alias{process_encoder}
 \title{Process serialize function for different languages}
 \usage{
-process_serialize_function(
-  serialize_function,
-  lang,
-  parent_env = parent.frame()
-)
+process_encoder(encoder, lang, parent_env = parent.frame())
 }
 \arguments{
-\item{serialize_function}{Language-specific serializer (see above)}
+\item{encoder}{Language-specific serializer (see above)}
 
 \item{lang}{Language string}
 

[…]
