### 2025-09-06 b6431ee
encoding/decoding vignette
