+++ pkgdown.yml
@@ -6,6 +6,7 @@ articles:
   cmdstanr: cmdstanr.html
   core-functions: core-functions.html
   debugging: debugging.html
+  encoding-decoding: encoding-decoding.html
   gc: gc.html
   importing-data: importing-data.html
   intro-concepts: intro-concepts.html
@@ -13,7 +14,7 @@ articles:
   polyglot: polyglot.html
   scope: scope.html
   tutorial: tutorial.html
-last_built: 2025-09-06T18:59Z
+last_built: 2025-09-06T19:29Z
 urls:
   reference: https://b-rodrigues.github.io/rixpress/reference
   article: https://b-rodrigues.github.io/rixpress/articles
