+++ data-raw/py2r_example/gen-pipeline.R
@@ -10,7 +10,7 @@ d1 <- rxp_py(
   # reticulate doesn't support polars DFs yet, so need to convert
   # first to pandas DF
   name = mtcars_pl_am,
-  py_expr = "mtcars_pl.filter(polars.col('am') == 1).to_pandas()",
+  expr = "mtcars_pl.filter(polars.col('am') == 1).to_pandas()",
 )
 
 d2 <- rxp_py2r(
@@ -31,7 +31,7 @@ d3_1 <- rxp_r2py(
 
 d4 <- rxp_py(
   name = mtcars_tail_py,
-  py_expr = 'mtcars_head_py.tail()',
+  expr = 'mtcars_head_py.tail()',
 )
 
 d4_1 <- rxp_py2r(
