### 2025-09-09 b81bfdd
Change log file format to JSON
Updated log filename to use .json extension and changed the saving method from RDS to JSON format.
