### 2025-09-09 956d312
Change log file format from RDS to JSON
