### 2025-09-16 #706 @ropensci-review-bot
Approved! Thanks @b-rodrigues for submitting and @wlandau, @amart90 for your reviews! :grin:
